% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.flowSet.SE.R
\name{as.flowSet.SE}
\alias{as.flowSet.SE}
\title{Transform a \code{fcs.SCE} object into a \code{flowset} object}
\usage{
as.flowSet.SE(fcs.SCE, assay.i)
}
\description{
Transform a \code{fcs.SCE} object into a \code{flowset} object
}
