% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.removal.R
\name{batch.removal}
\alias{batch.removal}
\title{Remove batch effect}
\usage{
  batch.removal(fcs.SCE, assay.i = "transformed", method, batch,
  new.matrix.name = "normalized", nfeatures = 10, dims = 1:10,
  threads = NULL)

}
\arguments{
  \item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

  \item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"transformed"}.}

  \item{method}{Methodology to perform batch correction. Possible values are "CCA" or "harmony".}

  \item{batch}{Variable name from \code{colData(FCS.SCE)} object to remove batch effect.}

  \item{new.matrix.name}{New normalized matrix name (it will stored within the \code{FCS.SE} object). Default = \code{"normalized"}.}

  \item{nfeatures}{Number of features (markers) to select as top variable features. Default = \code{10}.}

  \item{dims}{Number of dimensions to calculated with CCA's method. Default = \code{1:10}.}

  \item{threads}{Number of threads for multithreading. Default = \code{NULL} (i.e., all cores minus one).}
}
\description{
  It removes posible batch effects present in multiple FCS stored in a \code{FCS.SCE} object. It based on: \enumerate{
    \item Seurat's Canonical Correlation Analysis (CCA), see the \href{https://satijalab.org/seurat/Seurat_AlignmentTutorial.html}{tutorial} for more information.
    \item \href{https://portals.broadinstitute.org/harmony/}{Harmony}'s methodology.}
}
\examples{
\dontrun{
fcs <- batch.removal(fcs, method = "harmony", batch = "patient_id", new.matrix.name = "harmony")
fcs <- batch.removal(fcs, method = "cca", batch = "patient_id", new.matrix.name = "cca")
}
}
\keyword{alignment}
\keyword{dataset}
\keyword{canonical}
\keyword{correlation}
\keyword{normalization}
\keyword{removal}
\keyword{effect}
\keyword{batch}
