% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.cell.clustering.R
\name{boxplot.cell.clustering}
\alias{boxplot.cell.clustering}
\title{Boxplots for identified clusters}
\usage{
boxplot.cell.clustering(
  fcs.SCE,
  assay.i = "normalized",
  cell.clusters,
  condition = "condition",
  return.mode = "percentage",
  log.trans = FALSE,
  color.by = condition,
  facet = FALSE,
  facet.free.scale = "free_x",
  y.limits,
  return.stats = FALSE,
  plot.only.sig = c(F, 0.05),
  cond.colors,
  color.by.colors
)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT:fcs.SCE]{FlowCT::fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{cell.clusters}{Name of column containing clusters identified through \code{\link[FlowCT:clustering.flow]{FlowCT::clustering.flow()}}.}

\item{condition}{Column name from the \code{colData(fcs.SCE)} object which contains condition information.}

\item{return.mode}{String for specifying if final resuls should be proportions ("percentage") or raw counts ("counts"). Default = \code{"percentage"}.}
  
\item{log.trans}{Logarithmic transformation of counts/percentage values?. Default = \code{FALSE}.}
  
\item{color.by}{Variable name (from \code{colData(fcs.SCE)}) for lines coloring. Default, same as \code{condition}.}

\item{facet}{Logical indicating if splitting boxplots by cell clusters. Default = \code{FALSE}.}

\item{facet.free.scale}{If \code{facet = TRUE}, string indicating how scales would be shared across all facets. Possible values: \code{"free_x"} (default), \code{"free_y"} and \code{"free"}.}

\item{shape.by}{Variable name (from \code{colData(fcs.SCE)}) for dot shaping. Default = \code{NULL}.}

\item{y.limits}{Numeric vector with limits for y-axis (minimum, maximum).}

\item{return.stats}{Logical indicating if calculated statistics should be returned in a new variable. Default = \code{FALSE}.}

\item{plot.only.sig}{Vector indicating if only significant cell clusters should be displayed (logical element) and the P-value cutoff for selecting those ones (numerical element). Default = \code{c(FALSE, 0.05)}.}

\item{cond.colors}{Vector with colors for plotting option \code{condition}.}
  
\item{color.by.colors}{Vector with colors for plotting option \code{color.by}.}
}
\description{
It draws a boxplot with cell clusters identified through \code{\link[FlowCT:fsom.clustering]{FlowCT::fsom.clustering()}}.
}
\examples{
\dontrun{
# option 1: show all cell clusters and return statistics
bx_sig <- boxplot.cell.clustering(fcs.SCE = fcs, cell.clusters = "SOM_named", facet = T,
    facet.free.scale = "free", return.stats = T)

# option 2: show only those significant cell clusters
boxplot.cell.clustering(fcs.SCE = fcs, cell.clusters = "SOM_named",
    plot.only.sig = c(T, 0.1))
}
}
\keyword{boxplot}
\keyword{cell}
\keyword{clusters}
\keyword{differential}
\keyword{distributions}
