% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cell.count.bx}
\alias{cell.count.bx}
\title{cell.count.bx}
\usage{
cell.count.bx(fcs.SE, assay.i = "normalized", x.axis,
  color.by = x.axis, label.by = NULL, limits = NULL)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{assay.i}{Name of matrix stored in the \code{FCS.SE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{x.axis}{Variable name (from \code{colData(fcs.SE)}) for drawing the x-axis in plot.}

\item{color.by}{Variable name (from \code{colData(fcs.SE)}) for coloring the plot. Default = \code{x.axis} (i.e., the same specfied in \code{x.axis}).}

\item{label.by}{Variable from (from \code{colData(fcs.SE)}) for labeling. Default = \code{NULL}.}

\item{limits}{Numeric vector with limits for plotting (minimum, maximum). Default = \code{NULL} (i.e., automatically calculated from data).}
}
\description{
This function draws a boxplot according to the cell count for a specified condition.
}
\examples{
\dontrun{cell.count.bx(fcs_seL, assay.i = "normalized", x.axis = "condition")}
}
\keyword{boxplot}
\keyword{cell}
\keyword{count}
