% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{circ.tree}
\alias{circ.tree}
\title{circ.tree}
\usage{
circ.tree(fcs.SE, assay.i = "normalized", cell.clusters,
  dist.method = "euclidean", hclust.method = "average",
  nodes = "display", open.angle = 100, dendro.labels = FALSE,
  scale.size = 10)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{assay.i}{Name of matrix stored in the \code{FCS.SE} object from which drawing the circular tree. Default = \code{"normalized"}.}

\item{cell.clusters}{A vector with clusters identified through \code{\link[FlowCT:fsom.clustering]{FlowCT::fsom.clustering()}} (and, normaly, later renamed).}

\item{dist.method}{Distance method measurement to be used. Possible values are "euclidean" (default), "maximum", "manhattan", "canberra", "binary" or "minkowski".}

\item{hclust.method}{Hierarchical clustering method to be used. Possible values are "average" (default), "ward.D", "ward.D2", "single", "complete", "mcquitty", "median" or "centroid".}

\item{nodes}{If \code{"display"} (default), nodes will be numered. If contains a numeric vector with node numbers, areas defined from these nodes will be differently colored.}

\item{open.angle}{Angle aperture circular layout. Default = \code{100}.}

\item{dendro.labels}{Logical whether adding labels to dendrogram. Default = \code{FALSE}.}

\item{scale.size}{Numerical value indicating how much scale points in the dendogram terminal nodes. Default = \code{10}.}
}
\description{
This function plots a circular dendrogram and a heatmap according a \code{FCS.SE}. Every leaf has a different colored point size regarding the cell type and the frequency for each cluster.
For additional information go to \href{https://guangchuangyu.github.io/software/ggtree/documentation/}{\code{\link{ggtree}}} package.
}
\examples{
\dontrun{
# step 1: display all node numbers to select how to coloring areas
circ.tree(fcs.SE = fcs_seL, cell.clusters = fcs_seL$SOM_L_named, nodes = "display")

# step 2: color areas indicating node numbers
circ.tree(fcs.SE = fcs_seL, cell.clusters = fcs_seL$SOM_L_named, nodes = c(17, 25))
}
}
\keyword{circular}
\keyword{clustering}
\keyword{dendrogram}
\keyword{hierachical}
\keyword{nodes}
\keyword{tree}
