% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{circ.tree.selectNodes}
\alias{circ.tree.selectNodes}
\title{circ.tree.selectNodes}
\usage{
circ.tree.selectNodes(expr, expr_saturated, cell_clusters,
  dist_method = "euclidean", hclust_method = "average")
}
\arguments{
\item{expr}{A data.frame with marker expression values for each cell}

\item{expr_saturated}{A data.frame with marker expression values for each cell where those expression values greater or smaller that 1 and 0, respectively, must be coherced to 1 and 0.}

\item{cell_clusters}{A vector with clusters identified in a previous clustering method.}

\item{dist_method}{The distance measure to be used. Possible values are "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".}

\item{hclust_method}{The agglomeration method to be used. Possible values are "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median" or "centroid".}
}
\description{
This function plots a circular dendrogram for clusters previously calculated in order to select those nodes for coloring in \code{\link[FlowCT:circ.tree]{FlowCT::circ.tree()}}.
For additional information go to \href{https://guangchuangyu.github.io/software/ggtree/documentation/}{\code{\link{ggtree}}} package.
}
\examples{
\dontrun{
circ.tree.selectNodes(expr = matrix_clusters, expr_saturated = matrix_clusters01,
    cell_clusters = cell_clustering)}
}
\keyword{circular}
\keyword{dendrogram}
\keyword{node}
\keyword{tree}
