% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cluster.heatmap}
\alias{cluster.heatmap}
\title{plot_clustering_heatmap_wrapper}
\usage{
cluster.heatmap(expr, expr_saturated, cell_clusters)
}
\arguments{
\item{expr}{A data.frame with marker expression values for each cell}

\item{expr_saturated}{A data.frame with marker expression values for each cell where those expression values greater or smaller that 1 and 0, respectively, must be coherced to 1 and 0.}

\item{cell_clusters}{A vector with clusters identified in a previous clustering method.}
}
\description{
This function draws a heatmap collapsing identified clusters and showing cell percentaje for each one.
}
\examples{
\dontrun{
plot_clustering_heatmap_wrapper(expr = expr[,surface_markers],
    expr_saturated = expr01[,surface_markers], cell_clusters = df$FlowSOM)}
}
\keyword{cell}
\keyword{clusters}
\keyword{heatmap}
\keyword{percentajes}
