% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.flow.R
\name{clustering.flow}
\alias{clustering.flow}
\title{Clustering FCS data}
\usage{
  clustering.flow(fcs.SCE, assay.i = "normalized", method, scale = FALSE,
  markers.to.use = "all", num.k, seurat.res = 0.4, seurat.dims = 1:10)

}
\arguments{
  \item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT.v2:fcs.SCE]{FlowCT.v2::fcs.SCE()}}.}

  \item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate SOM clustering. Default = \code{"normalized"}.}

  \item{method}{What method should be used for clustering purposes. Available ones are "SOM", "Phenograph", "Seurat" and "PARC".}

  \item{scale}{Should be data be scale before SOM clustering? (only available for this method). Default = \code{"FALSE"}.}

  \item{markers.to.use}{Markers used for considering within the clustering calculation. Default = \code{"all"}.}

  \item{num.k}{Number of clusters to calculate for methods "SOM" and "Phenograph".}

  \item{seurat.res}{Seurat's resolution to calculate clustering (it indicates the ‘granularity’ of the downstream clustering, with increased values leading to a greater number of clusters). Default = \code{0.4}.}

  \item{seurat.dims}{Number of dimensions to calculated with Seurat's method. Default = \code{1:10}.}
}

\description{
  This function offers multiple automatic clustering methods for a \code{FCS.SCE} object. Those methods available are: \enumerate{
  \item SOM: Self-Organizing Map (SOM) based on \href{https://github.com/SofieVG/FlowSOM}{\code{FlowSOM} package}.

  \item Phenograph: An implementation of PhenoGraph algorithm (\href{https://github.com/JinmiaoChenLab/Rphenograph}{here} for more information).

  \item Seurat: Clustering method based on \href{https://satijalab.org/seurat/}{Seurat}'s single-cell analysis.

  \item PARC: R implementation for Phenotyping by Accelerated Refined Community-partitioning (PARC) method, see \href{https://github.com/ShobiStassen/PARC}{Python module}.
  }
}
\examples{
  \dontrun{
	fcs <- clustering.flow(fcs, method = "som", num.k = 20)
	fcs <- clustering.flow(fcs, method = "phenograph", num.k = 40)
	fcs <- clustering.flow(fcs, method = "Seurat", seurat.res = 0.5)
	fcs <- clustering.flow(fcs, method = "parc", assay.i = "transformed")
  }
}
\keyword{SOM}
\keyword{unsupervised}
\keyword{clustering}
\keyword{Seurat}
\keyword{PARC}
\keyword{Phenograph}
