% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.rename.R
\name{clusters.rename}
\alias{clusters.rename}
\title{It renames the numerical clusters detected through \code{\link[FlowCT]{fsom.clustering()}} to specific cell populations identified by flow cytometry user.}
\usage{
clusters.rename(x, cluster, name)
}
\arguments{
\item{x}{Vector with numeric values corresponding to clusters detected with \code{\link[FlowCT]{fsom.clustering()}}.}

\item{cluster}{Numeric vector or column name with numeric values for replacing.}

\item{name}{Cell population names to replace numeric values from \code{cluster} column.}
}
\description{
It renames the numerical clusters detected through \code{\link[FlowCT]{fsom.clustering()}} to specific cell populations identified by flow cytometry user.
}
\examples{
\dontrun{
# option 1  
head(replace_data)
 #   original_cluster new_cluster
 # 1 1                debris     
 # 2 2                lymphocytes
 # 3 3                monocytes
 # 4 4                lymphocytes
 # 5 5                eosinophils
 # 6 6                eosinophils
fcs$SOM_named <- clusters.rename(fcs$SOM, cluster = replacedata$original_cluster, name = replacedata$new_cluster)

# option 2
fcs$SOM_named <- clusters.rename(fcs$SOM, cluster = 1:6, name = c("debris", "lymphocytes", "monocytes", "lymphocytes", "eosinophils", "eosinophils"))
}
}
\keyword{assignment}
\keyword{clusters}
\keyword{population}
\keyword{renaming}
