% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.annot.pheatmap.R
\name{col.annot.pheatmap}
\alias{col.annot.pheatmap}
\title{Create a palette of colors suitable for doing a heatmap}
\usage{
col.annot.pheatmap(metadata, colors = NULL)
}
\description{
Create a palette of colors suitable for doing a heatmap
}
