% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine.subclusterings}
\alias{combine.subclusterings}
\title{It combines initial \code{FCS.SE} object (without subclustering) with other \code{FCS.SE} objects with subclustering analysis coming from downstream steps and generates a new \code{FCS.SE} object. This final \code{FCS.SE} object has an additional column combining all information from initial and subclustering analysis.}
\usage{
combine.subclusterings(initial.fcs.SE, subclustering.fcs.SE,
  clusters.named = "SOM_named")
}
\arguments{
\item{initial.fcs.SE}{A \code{FCS.SE} object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}. The initial one, without extracting any cell population.}

\item{subclustering.fcs.SE}{A list with all \code{FCS.SE} object(s) generated in the subclustering analysis (they have to come from the original \code{initial.fcs.SE}).}

\item{clusters.named}{Column name from the \code{initial.fcs.SE} object which contains renamed clusters (through \code{\link[FlowCT:clusters.rename]{FlowCT::clusters.rename()}}) and has been used to extract cell populations for subclustering steps.}
}
\description{
It combines initial \code{FCS.SE} object (without subclustering) with other \code{FCS.SE} objects with subclustering analysis coming from downstream steps and generates a new \code{FCS.SE} object. This final \code{FCS.SE} object has an additional column combining all information from initial and subclustering analysis.
}
\examples{
\dontrun{
 fcs_se_final <- combine.subclusterings(initial.fcs.SE = fcs_se, clusters.named = "SOM_named", 
    subclustering.fcs.SE = list(fcs_se_lymphos, fcs_se_monos))
}
}
\keyword{FCS.SE}
\keyword{combine}
\keyword{final}
\keyword{object}
\keyword{subclustering}
