% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cor.plot.conditions}
\alias{cor.plot.conditions}
\title{cor.plot.conditions}
\usage{
cor.plot.conditions(data, colname_condition, colname_patientID, conditions,
  metadata)
}
\arguments{
\item{data}{A data.frame with proportions for each cell type.}

\item{colname_condition}{Colname of the column containing differential condition.}

\item{colname_patientID}{Colname of the column containing sample (or patient) identification.}

\item{conditions}{A vector with those conditions within colname_condition.}

\item{metadata}{A data.frame with additional information for cell proportions (it must include colname_condition and colname_patientID)}
}
\description{
This function draws a correlation plot (based in \code{\link[corrplot:corrplot]{corrplot::corrplot()}}) between specified conditions.
For additional information go to \href{https://cran.r-project.org/web/packages/corrplot/vignettes/corrplot-intro.html}{\code{\link{corrplot}}} package.
}
\examples{
\dontrun{
cor.plot.conditions(data = dataset_final, colname_condition = "condition",
    colname_patientID = "patient_id", conditions = c("PB", "BM"), metadata = md)}
}
\keyword{corr}
\keyword{correlation}
\keyword{plot}
