% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxregs.R
\name{cox.plot}
\alias{cox.plot}
\title{Forest plot for uni/multivariable Cox regressions}
\usage{
cox.plot(
  pop.cutoff.obj,
  time.var,
  event.var,
  cox.type,
  variables,
  ref.var,
  xlim = c(0, 10),
  return.stats = T,
  log.scale = F,
  hide.nosig = F
)
}
\arguments{
\item{pop.cutoff.obj}{An object generated through \code{\link[FlowCT.v2:pop.cutoff]{FlowCT.v2::pop.cutoff()}}.}

\item{time.var}{Survival time variable.}

\item{event.var}{Variable with event censoring.}

\item{cox.type}{Cox regression type: "unicox" or "multicox".}

\item{variables}{Vector with variables for performing the regression. If nothing is detailed (\code{NULL}, default), all immune's cutoffs will be used.}

\item{ref.var}{Variable for being used as reference. Default, alphabetical order.}

\item{xlim}{Limits for x-axis plotting. Default = \code{c(0,10)}.}

\item{return.stats}{Logical indicating if calculated statistics must be returned. Default = \code{TRUE}.}

\item{log.scale}{Should x-axis be log-transformed (log10). Default = \code{FALSE}.}

\item{hide.nosig}{Hide non-significant cell populations. Default = \code{FALSE}.}
}
\description{
Plotting (Forest plot) of uni/multivariable Cox regressions.
}
\examples{
\dontrun{
unicox <- cox.plot(pop.cutoff.obj = pop_cuts, time.var, event.var, cox.type = "multicox", ref.var = "low")
}
}
\keyword{Cox}
\keyword{forest-plot}
\keyword{multivariable}
\keyword{regression}
\keyword{univariable}
