% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdots.cell.clustering.R
\name{diffdots.cell.clustering}
\alias{diffdots.cell.clustering}
\title{Longitudinal differential dotplot}
\usage{
diffdots.cell.clustering(
  fcs.SCE,
  assay.i = "normalized",
  cell.clusters,
  condition.column,
  psig.cutoff = 0.05,
  return.stats = T,
  colors = NULL,
  return.mode = "percentage",
  hide.nosig = F
)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT:fcs.SCE]{FlowCT::fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{cell.clusters}{A vector with clusters identified through \code{\link[FlowCT:fsom.clustering]{FlowCT::fsom.clustering()}} (and, normaly, later renamed).}

\item{condition.column}{Column name from the \code{colData(fcs.SCE)} object which contains condition information. De}

\item{psig.cutoff}{P-value cutoff. Default = \code{0.05}.}

\item{return.stats}{Logical indicating if calculated statistics should be returned in a new variable. Default = \code{TRUE}.}

\item{colors}{Vector with colors for plotting. Default = \code{NULL} (i.e., it will choose automatically a vector of colors according to \code{\link[FlowCT:div.colors]{FlowCT::div.colors()}}).}

\item{hide.nosig}{Logical indicating whether hiding non-significal cell populations. Default = \code{FALSE}.}
}
\description{
It draws a differential dot plot (longitudinaly) according condition for each cell cluster identified.
}
\examples{
\dontrun{
diffDots.cell.clustering(fcs.SCE = fcs, cell.clusters = fcs$SOM_named, return.stats = F)
}
}
\keyword{Dumbbell}
\keyword{differential}
\keyword{dotplot}
\keyword{longitudinal}
\keyword{plot}
