% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dotplot.DE}
\alias{dotplot.DE}
\title{It draws dotplot for each cell cluster identified and each marker to facilitate identification of cell populations in each cell cluster.}
\usage{
dotplot.DE(fcs.SE, assay.i = "normalized",
  clusters.named = "SOM_named", markers.to.use = "all",
  psig.cutoff = 0.05, return.stats = F, scale.size = 9)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{assay.i}{Name of matrix stored in the \code{FCS.SE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{clusters.named}{Column name from the \code{initial.fcs.SE} object which contains renamed clusters (through \code{\link[FlowCT:clusters.rename]{FlowCT::clusters.rename()}}).}

\item{markers.to.use}{Vector with markers to use. Default = \code{"all"}.}

\item{psig.cutoff}{P-value cutoff. Default = \code{0.05}.}

\item{return.stats}{Logical indicating if calculated statistics should be returned in a new variable. Default = \code{FALSE}.}

\item{scale.size}{Numerical value indicating how much scale points. Default = \code{9}.}
}
\description{
It draws dotplot for each cell cluster identified and each marker to facilitate identification of cell populations in each cell cluster.
}
\examples{
\dontrun{
de.dotplot(fcs.SE = fcs_se, markers.to.use = c("CD8", "CD27", "CCR4", "CD45RA", "CD4"), 
    clusters.named = "SOM_named")
}
}
\keyword{cell}
\keyword{cluster}
\keyword{differential}
\keyword{dotplot}
\keyword{identification}
\keyword{markers}
