% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dr.plotting}
\alias{dr.plotting}
\title{dr.plotting}
\usage{
dr.plotting(dr, dr.calculated, n.dims = c(1, 2),
  color.by = "expression", facet.by = NULL, omit.markers = NULL,
  title = "", label.by = NULL, size = 2, raster = c(T, 1000))
}
\arguments{
\item{dr}{A object with DR generated with \code{\link[FlowCT:dim.reduction]{FlowCT::dim.reduction()}} or a \code{data.frame} with DR, expression and metadata information (like the first element list of the object generated with \code{\link[FlowCT:dim.reduction]{FlowCT::dim.reduction()}}).}

\item{dr.calculated}{String indicating the desired DR to plot (this indicated DR should be prevoulsy calculated to being plotted).}

\item{n.dims}{Vector indicating the two DR components to plot. Default = \code{c(1,2)}.}

\item{color.by}{Variable from (from \code{colData(fcs.SE)}) for dots coloring. If \code{color.by = "expression"} (default), plot will be splitted for each marker (\code{facet.by}).}

\item{facet.by}{Variable from (from \code{colData(fcs.SE)}) for plot spliting. Default = \code{NULL}.}

\item{omit.markers}{Vector with markers to omit when plotting with \code{color.by = "expression"}. Default = \code{NULL}.}

\item{title}{Title to add to the plot.}

\item{label.by}{Variable from (from \code{colData(fcs.SE)}) for dots labeling. Default = \code{NULL}.}

\item{size}{Point size. Default = \code{0.5}.}

\item{raster}{Vector indicating if image should be rasterized (logical element) and the number of pixels to consider (numerical element). It is based on (\href{https://github.com/exaexa/scattermore}{scattermore package}). Default = \code{c(T, 1000)}.}
}
\description{
This function plots the indicated dimensional reduction (DR) from a previously calculated \code{\link[FlowCT:dim.reduction]{FlowCT::dim.reduction()}} object.
}
\examples{
\dontrun{
dr.plotting(dr, dr.calculated = "tSNE", color.by = "condition")
dr.plotting(dr, dr.calculated = "UMAP", color.by = "patient_id")
dr.plotting(dr, dr.calculated = "PCA", color.by = "SOM", facet.by = "condition")
}
}
\keyword{PCA}
\keyword{UMAP}
\keyword{dimensional}
\keyword{plotting}
\keyword{reduction}
\keyword{tSNE}
