% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumbPlot.R
\name{dumbPlot}
\alias{dumbPlot}
\title{Dumbbell plot with clusters}
\usage{
dumbPlot.cell.clustering(
  fcs.SCE,
  assay.i = "normalized",
  cell.clusters,
  condition.column,
  psig.cutoff = 0.05,
  return.stats = F,
  colors
)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT:fcs.SCE]{FlowCT::fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{cell.clusters}{A vector with clusters identified through \code{\link[FlowCT:fsom.clustering]{FlowCT::fsom.clustering()}} (and, normaly, later renamed).}

\item{condition}{Column name from the \code{colData(fcs.SCE)} object which contains condition information.}

\item{psig.cutoff}{P-value cutoff. Default = \code{0.05}.}

\item{return.stats}{Logical indicating if calculated statistics should be returned in a new variable. Default = \code{FALSE}.}

\item{colors}{Vector with colors for plotting.}
}
\description{
It draws a Dumbbell plot according condition for each cell cluster identified.
}
\examples{
\dontrun{
diffDots.cell.clustering(fcs.SCE = fcs, cell.clusters = "SOM_named")
}
}
\keyword{Dumbbell}
\keyword{differential}
\keyword{dotplot}
\keyword{longitudinal}
\keyword{plot}
