% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.metaFCS.R
\name{export.metaFCS}
\alias{export.metaFCS}
\title{export.metaFCS}
\usage{
export.metaFCS(fcs.SCE, assay.i = "raw", output.name = NULL,
  output.folder = getwd(), separate.fcs = FALSE,
  output.suffix = "meta")
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT]{fcs.SCE()}}.}

\item{output.name}{Name for generated FCS file. If \code{NULL} (default), the project name specified in the \code{fcs.SCE} building will be used.}

\item{output.folder}{Name of the folder within store the new generated FCS files. Default = \code{getwd()} (i.e., current directory).}

\item{separate.fcs}{Logical indicating whether FCS must be exporting separatelly or not. Default = \code{FALSE}.}

\item{output.suffix}{Suffix to be added to the new generated FCS files. Default = \code{"meta"}.}

\item{dr.object}{Object created with \code{\link[FlowCT]{dim.reduction()}} function or a table combining DR and experimental metadata information.}
}
\description{
It creates a FCS file containing all analysis incorpored to colData(fcs.SCE) as well as the dimensional reduction coordinates calculated.
}
\examples{
\dontrun{
# option 1: save entire analysis as a single FCS file
export.metaFCS(fcs.SCE = fcs, output.name = "final_analysis")

# option 2: save separatelly each file contained within the fcs.SCE object
export.metaFCS(fcs.SCE = fcs, separate.fcs = T, output.suffix = "meta.final")
}
}
\keyword{FCS}
\keyword{exporting}
\keyword{final}
\keyword{generation}
