% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs.SCE.R
\name{fcs.SCE}
\alias{fcs.SCE}
\title{Read multiple FCS files and create a \code{fcs.SCE} object}
\usage{
fcs.SCE(
  filelist = NULL,
  directory = getwd(),
  pattern = ".fcs$",
  events = "all",
  dataset = 1,
  num.threads = NULL,
  metadata,
  transformation = "arcsinh",
  transf.cofactor = 500,
  project.name = "noname"
)
}
\arguments{
\item{filelist}{A vector with full path of FCS files to be read, commonly generated through \code{\link[base:list.files]{base::list.files()}}. If \code{NULL}, this file list will be generated as indicated below.}

\item{directory}{If \code{filelist = NULL}, those files stored in this location will be read. Default = \code{getwd()} (current directory).}

\item{pattern}{Pattern for reading files within \code{directory}. Default = \code{"fcs"}.}

\item{events}{Numeric vector indicating how many events to read in each FCS file (same behaviour as \code{\link[flowCore:read.FCS]{flowCcre::read.FCS()}}), those are the number of events for new and unified generated FCS files. Default = \code{"all"}.}

\item{dataset}{Specify data segment to be read in the FCS file (same behaviour as \code{\link[flowCore:read.FCS]{flowCcre::read.FCS()}}). Default = \code{1}.}

\item{num.threads}{Number of threads for reading the FCS files. For Windows users, you should to install \code{parallelsugar} from \href{https:https://github.com/nathanvan/parallelsugar}{Nathanvan's GitHub}. Defult = \code{NULL}, i.e., not parallelization.}

\item{metadata}{Metadata information for each FCS files included in the analysis. It will be used to assign to each event an specific metadata and it will be added to \code{metadata(fcs.SCE)} as \code{reduced_metadata}.}

\item{transformation}{Because FCS data is normally exported in a logarithmic format, it is necessary its transformation for downstream analysis. Default (and the only one) = \code{arcsinh}.}

\item{transf.cofactor}{Cofactor numeric value for that \code{arcsinh} transformation. Values can ranged from 15 to 50 (for mass cytometry) and from 150 to 10,000 (for flow cytometry). Default = \code{500}.}

\item{project.name}{String naming the \code{fcs.SCE} generated. Default = \code{"noname"}.}
}
\value{

}
\description{
It reads and creates a \code{fcs.SCE} object (based on the \href{https://www.bioconductor.org/packages/release/bioc/html/SummarizedExperiment.html}{\code{SummarizedExperiment} class}) from FCS files in a specific folder or indicated in a vector. Important note: all files should have an identicar header (i.e., same name and markers order), if not only those common will be read (you can use \code{\link[FlowCT.v2:unify.FCSheaders]{FlowCT.v2::unify.FCSheaders()}} for doing that).
}
\examples{
\dontrun{
head(metadata_user)
 #       filename sample_id condition patient_id
 # 1 BM_017294.fcs         1        BM     017294
 # 2 BM_017564.fcs         2        BM     017564
 # 3 BM_017612.fcs         3        BM     017612
 # 4 PB_017294.fcs         4        PB     017294
 # 5 PB_017564.fcs         5        PB     017564
 # 6 PB_017612.fcs         6        PB     017612
fcs <- fcs.SCE(directory = "../data/", which.lines = 1000, metadata = metadata_user, transf.cofactor = 500)
}
}
\keyword{FCS}
\keyword{fcs.SCE}
\keyword{metadata}
\keyword{object}
\keyword{read}
\keyword{single-cell}
