% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduction_and_clustering.R
\name{fsom.clustering}
\alias{fsom.clustering}
\title{fsom.clustering}
\usage{
fsom.clustering(data, markers_to_use = "surface_markers",
  markers_to_plot = "tree", set.seed = 333, saveRDS = F)
}
\arguments{
\item{data}{A data.frame with the marker expression values for each cell.}

\item{markers_to_use}{Markers to use in the clustering calculation. Default = "surface_markers"}

\item{markers_to_plot}{Markers to plot in the final image. Default = "tree" (it plots a simple tree, without markers).}

\item{set.seed}{\code{\link[base:set.seed]{set.seed()}} for clustering calculation. Default = 333}

\item{saveRDS}{Logical whether saving the final calculated object. Default = \code{FALSE}}
}
\description{
This function calculates a SOM (Self-Organizing Map) clustering from a data.frame with marker expression values for all cells.
For additional information go to \href{https://github.com/SofieVG/FlowSOM}{\code{\link{FlowSOM}}} package.
}
\examples{
\dontrun{
fsom <- fsom.clustering(fcs, markers_to_use = "surface_markers",
    markers_to_plot = "tree")
fsom <- fsom.clustering(cell_expressions, markers_to_use = "surface_markers",
    markers_to_plot = "CXCR3")
    }
}
\keyword{SOM}
\keyword{clustering}
