% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReduction_and_clustering.R
\name{fsom.metaclustering}
\alias{fsom.metaclustering}
\title{fsom.metaclustering}
\usage{
fsom.metaclustering(fsom, num_clusters_metaclustering = 40,
  plotting = T, folder_name = "consensus_plots", set.seed = 333,
  saveRDS = F)
}
\arguments{
\item{fsom}{Object previously calculated with \code{\link[FlowCT:fsom.clustering]{FlowCT::fsom.clustering()}}.}

\item{num_clusters_metaclustering}{Number of clusters to reduce. Default = 40}

\item{plotting}{Logical whether plotting final SOM recalculation. Default = \code{TRUE}}

\item{folder_name}{Name for created folder where consensus clustering will be stored. Default = "consensus_plot"}

\item{set.seed}{\code{\link[base:set.seed]{set.seed()}} for clustering calculation. Default = 333}

\item{saveRDS}{Logical whether saving final calculated object. Default = \code{FALSE}}
}
\description{
This function calculates a consensus clustering to reduce the number of previous identified clusters through SOM (Self-Organizing Map) clustering.
For additional information go to \href{http://bioconductor.org/packages/release/bioc/vignettes/ConsensusClusterPlus/inst/doc/ConsensusClusterPlus.pdf}{\code{\link{ConsensusClusterPlus}}} package.
}
\examples{
\dontrun{metaclusters <- fsom.metaclustering(fsom = fsom, num_clusters_metaclustering = 40)}
}
\keyword{SOM}
\keyword{clustering}
\keyword{consensus}
\keyword{metaclustering}
