% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gauss.norm}
\alias{gauss.norm}
\title{gauss.norm}
\usage{
gauss.norm(fcs.SE, marker.to.norm)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{marker.to.norm}{Marker name(s) to normalize (based on visual inspection with \code{\link[FlowCT:multidensity]{FlowCT::multidensity()}}).}

\item{norm.matrix.name}{New normalized matrix name (it will stored within the \code{FCS.SE} object). Default = \code{"normalized"}.}
}
\description{
It performs a Gaussian normalization based on \code{\link[flowStats:gaussNorm]{flowStats::gaussNorm()}} function.
}
\examples{
\dontrun{
 fcs_se <- gauss.norm(fcs.SE = fcs_se, marker.to.norm = c("CD62L", "CCR4", "SSC_A"))
}
}
\keyword{Gaussian}
\keyword{alignment}
\keyword{marker}
\keyword{normalization}
