% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lims.FCS}
\alias{lims.FCS}
\title{lims.FCS}
\usage{
lims.FCS(fcs)
}
\arguments{
\item{fcs}{A data.frame with marker expression values for each cell.}
}
\description{
This function estimates the minimum and maximum values for an object with multiple FCS files. Usefull for plotting limits.
}
\examples{
\dontrun{lims.FCS(fcs_raw)}
}
