% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{marker.names}
\alias{marker.names}
\title{It shows marker names of a \code{FCS.SE} object and renames them accordin a new vector provided by the user.}
\usage{
marker.names(fcs.SE, new.names = NULL)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{new.names}{Vector with new channel/marker names (it must has the same length that \code{FCS.SE}'s markers). Default = \code{NULL} (i.e., markers will not be renamed, only displayed).}
}
\description{
It shows marker names of a \code{FCS.SE} object and renames them accordin a new vector provided by the user.
}
\examples{
\dontrun{
marker.names(fcs_se)
new_names <- c("FSC_A", "FSC_H", "SSC_A", "SSC_H", "CD62L", "CXCR3", "CD8", "CCR4", "CCR6", "CD4", "CD45", "CD27")
fcs_se <- marker.names(fcs_se, new.names = new_names)
}
}
\keyword{marker}
\keyword{renaming}
\keyword{names}
