% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{markers.equal}
\alias{markers.equal}
\title{markers.equal}
\usage{
markers.equal(markers1, markers2)
}
\arguments{
\item{markers1}{First string with the marker names.}

\item{markers2}{Second string with the marker names.}
}
\description{
This function compares if two strings are equal (in this case, for marker comparison).
}
\examples{
\dontrun{markers.equal(surface_markers, fcs_raw@colnames)}
}
