% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{markers.names}
\alias{markers.names}
\title{markers.names}
\usage{
markers.names(fcs, new_names = NULL)
}
\arguments{
\item{fcs}{A \code{flowSet} object read with \code{\link[flowCore:read.flowSet]{flowCore::read.flowSet()}}.}

\item{new_names}{A string with new markers names to rename the \code{flowSet} object. Default = \code{NULL}}
}
\description{
This function shows default names for chanels in FCS files and offers the option to change them manually.
}
\examples{
\dontrun{
markers.names(fcs_raw) #it only shows marker names...
fcs_raw <- markers.names(fcs_raw, new_names = c("FSC_A", "FSC_H", "SSC_A", "SSC_H",
    "CD62L", "CXCR3", "CD8", "CCR4", "CCR6", "CD4", "CD27", "CD45RA"))} #it shows and changes marker names according supplied ones
}
