% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{median.dr}
\alias{median.dr}
\title{This function draws a plot from internally calculated dimensional reduction (DR).}
\usage{
\method{median}{dr}(fcs.SE, assay.i = "normalized",
  markers.to.use = "all", dr.method = "PCA", perplexity.tsne = 15,
  n.neighbors.umap = 50, color.by, label.by = NULL, size = 3)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{assay.i}{Name of matrix stored in the \code{FCS.SE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{markers.to.use}{Vector with markers to use. Default = \code{"all"}.}

\item{dr.method}{DR method to calculate. Possible values are "PCA", "tSNE" or "UMAP", individually, or "all". Take into account the number of samples for this DR, for few samples t-SNE and UMAP are useless.}

\item{perplexity.tsne}{Value for perplexity parameter in tSNE calculation (\href{https://distill.pub/2016/misread-tsne/}{more information}). For little number of samples this DR (t-SNE) cannot be calculated because preplexity restrictions. Default = \code{15}.}

\item{n.neighbors.umap}{Value for neighbors parameter in UMAP calculation (\href{https://www.math.upenn.edu/~jhansen/2018/05/04/UMAP/}{more information}). Default = \code{50}.}

\item{color.by}{Variable name (from \code{colData(fcs.SE)}) for lines coloring.}

\item{label.by}{Variable name (from \code{colData(fcs.SE)}) for labelling.}

\item{size}{Point size. Default = \code{3}.}
}
\description{
This function draws a plot from internally calculated dimensional reduction (DR).
}
\examples{
\dontrun{
median.dr(fcs_se, color.by = "condition")
}
}
\keyword{dimensional}
\keyword{expression}
\keyword{median}
\keyword{reduction}
\keyword{values}
