% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{median.values}
\alias{median.values}
\title{It calculates median values according a specfied variable, normally "filename".}
\usage{
\method{median}{values}(fcs.SE, assay.i = "raw", var = "filename")
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{assay.i}{Name of matrix stored in the \code{FCS.SE} object from which calculate medians. Default = \code{"raw"}.}

\item{var}{Variable for grouping and calculating medians. Default = \code{"filename"} (i.e., names of FCS files).}
}
\description{
It calculates median values according a specfied variable, normally "filename".
}
\examples{
\dontrun{
mfis_FCS <- median.values(fcs.SE = fcs_se)
med_SOM_clust <- median.values(fcs.SE = fcs_se, assay.i = "normalized", var = "SOM_named")
}
}
\keyword{MFI}
\keyword{fluorescence}
\keyword{intensity}
\keyword{median}
\keyword{values}
