% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxregs.R
\name{multicox}
\alias{multicox}
\title{Multivariable Cox Regression}
\usage{
multicox(
  data,
  covariates,
  pval.cutoff = 0.05,
  timevar,
  censvar,
  xlim = c(0, 100),
  log.scale = F,
  hide.nosig = T
)
}
\description{
Multivariable Cox Regression
}
