% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{multidensity}
\alias{multidensity}
\title{multidensity}
\usage{
multidensity(fcs.SE, assay.i, show.markers = "all", color.by,
  subsampling = NULL, interactive = F, ridgeline.lim = 15)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{assay.i}{Name of matrix stored in the \code{FCS.SE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{show.markers}{Vector with markers to plot. Default = \code{"all"}.}

\item{color.by}{Variable name (from \code{colData(fcs.SE)}) for lines coloring.}

\item{subsampling}{Numeric value indicating how many events use to calculate density lines and speed up plotting. Default = \code{NULL}.}

\item{interactive}{Logical indicating if the user can interact with the (only overlapping-lines) plot. Default = \code{FALSE}.}

\item{ridgeline.lim}{Numeric value specifying the limit for shifting from ridgeline-mode to overlapping-lines plot. Default = \code{15}.}
}
\description{
This function draws multidensity plot with all FCS files included in a \code{FCS.SE} object. If there are more files than limit specified in \code{ridgeline.lim}, instead of plottting density in a ridge-way all density lines will be overlapped.
}
\examples{
\dontrun{
multidensity(fcs.SE = fcs_se, assay.i = "normalized", subsampling = 1000)
multidensity(fcs_se, assay.i = 2, color.by = "file_name", ridgeline.lim = 0, 
    show.markers = c("CD62L", "CD4"), interactive = T)
}
}
\keyword{alignment}
\keyword{density}
\keyword{marker}
\keyword{normalization}
