% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.flw.R
\name{normalization.flw}
\alias{normalization.flw}
\title{Normalization for markers' expression}
\usage{
normalization.flw(
  fcs.SCE,
  assay.i = "transformed",
  marker.to.norm,
  norm.method,
  var.to.use,
  norm.matrix.name = "normalized"
)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT.v2:fcs.SCE]{FlowCT.v2::fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"transformed"}.}

\item{marker.to.norm}{Marker name(s) to normalize (based on visual inspection with \code{\link[FlowCT.v2:multidensity]{FlowCT.v2::multidensity()}}). Only applicable for methods "gauss" and "warp".}

\item{norm.method}{Methodology to perform each marker normalization. Possible values are "gauss" or "warp" (from \code{flowStats} package) or "harmony" (from \code{harmony} package for single-cell).}

\item{var.to.use}{Variable from \code{colData(fcs.SCE)} for calculating the normalization (only if \code{norm.method = "harmony"}).}

\item{norm.matrix.name}{New normalized matrix name (it will stored within the \code{fcs.SCE} object). Default = \code{"normalized"}.}
}
\description{
It normalizes expression abnormalities within selected markers or within the entire expression set.
}
\examples{
\dontrun{
 fcs_gauss <- normalization.flw(fcs.SCE = fcs, marker.to.norm = c("CD62L", "CCR4", "SSC_A"), norm.method = "gauss")
 fcs_warp <- normalization.flw(fcs.SCE = fcs, marker.to.norm = c("CD62L", "CCR4", "SSC_A"), norm.method = "warp")
 fcs_harmony <- normalization.flw(fcs.SCE = fcs, norm.method = "harmony", var.to.use = "condition")
}
}
\keyword{alignment}
\keyword{gaussian}
\keyword{harmony}
\keyword{marker}
\keyword{normalization}
\keyword{warp}
