% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_clustering_heatmap_wrapper}
\alias{plot_clustering_heatmap_wrapper}
\title{plot_clustering_heatmap_wrapper}
\usage{
plot_clustering_heatmap_wrapper(expr, expr_saturated, cell_clusters,
  cluster_merging = NULL)
}
\arguments{
\item{expr}{A data.frame with marker expression values for each cell}

\item{expr_saturated}{A data.frame with marker expression values for each cell where those expression values greater or smaller that 1 and 0, respectively, must be coherced to 1 and 0.}

\item{cell_clusters}{A vector with clusters identified in a previous clustering method.}

\item{cluster_merging}{???????????}
}
\description{
This function draws a heatmap taking account marker expression levels for each identified cluster and shows the percentaje of cells for each one.
}
\examples{
\dontrun{
plot_clustering_heatmap_wrapper(expr = expr[,surface_markers],
    expr_saturated = expr01[,surface_markers], cell_clusters = df$FlowSOM)}
}
\keyword{cell}
\keyword{clusters}
\keyword{heatmap}
\keyword{percentajes}
