% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{qc.and.removeDoublets}
\alias{qc.and.removeDoublets}
\title{qc.and.removeDoublets}
\usage{
qc.and.removeDoublets(directory = ".", reduction_computed = TRUE,
  reduction_suffix = "_red", output_suffix = "_preprocessed",
  output_folder = "results_HQsinglets")
}
\arguments{
\item{directory}{Directory with FCS files to process. Default = "." (current directory)}

\item{reduction_computed}{Logical indicating wheter previuous reduction has been computed (through \code{\link[FlowCT:reduce.FCS]{FlowCT::reduce.FCS()}}). Default = \code{TRUE}}

\item{reduction_suffix}{Suffix added previously to reduced files trough \code{\link[FlowCT:reduce.FCS]{FlowCT::reduce.FCS()}}.}

\item{output_suffix}{Suffix to be added to final FCS files. Default = "_preprocessed"}

\item{output_folder}{Folder name for storing final FCS files. Default = "results_HQsinglets"}
}
\description{
qc.and.removeDoublets
}
\examples{
\dontrun{qc.and.removeDoublets()}
}
\keyword{control}
\keyword{doublets}
\keyword{quality}
\keyword{remove}
