% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{reduce.FCS}
\alias{reduce.FCS}
\title{reduce.FCS}
\usage{
reduce.FCS(file_or_directory = ".", keep_n_events = 1e+05,
  output_folder = "reduced", output_suffix = "_red")
}
\arguments{
\item{file_or_directory}{FCS filename or directory where to search for FCS files. Default = "."}

\item{keep_n_events}{Number of events to keep in each FCS files. Default = 100000}

\item{output_folder}{Folder name for storing new reduced FCS files. Default = "reduced"}

\item{output_suffix}{Suffix to be added to reduced FCS files. Default = "_red"}
}
\description{
This function reduce the number of events for a single FCS file or multiple FCS files contained in a given folder.
}
\examples{
\dontrun{reduce.FCS()}
}
\keyword{events}
\keyword{reduce}
