% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove.pop}
\alias{remove.pop}
\title{It removes one or multiple cell populations from a \code{FCS.SE} object.}
\usage{
remove.pop(fcs.SE, population, clusters.named)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{population}{Name(s) of cell population(s) to be removed.}

\item{clusters.named}{Column name from the \code{colData(fcs.SE)} object which contains renamed clusters (through \code{\link[FlowCT:clusters.rename]{FlowCT::clusters.rename()}}).}
}
\description{
It removes one or multiple cell populations from a \code{FCS.SE} object.
}
\examples{
\dontrun{
fcs_se_rm <- remove.pop(fcs_se, clusters.named = "SOM_named", 
    population = c("debris", "unclassified"))
}
}
\keyword{cell}
\keyword{debris}
\keyword{population}
\keyword{remove}
