% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.tree.R
\name{rpart.tree}
\alias{rpart.tree}
\title{Recursive partitioning tree}
\usage{
rpart.tree(
  fcs.SCE,
  cell.clusters,
  variables,
  value = "percentage",
  time.var,
  event.var,
  xerror,
  rpart.control,
  return.data = T
)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT:fcs.SCE]{FlowCT::fcs.SCE()}}.}

\item{cell.clusters}{Name of column containing clusters identified through \code{\link[FlowCT:clustering.flow]{FlowCT::clustering.flow()}}.}

\item{variables}{Vector with variables for calculating the cutoff. If nothing is detailed (\code{NULL}, default), all immune populations from \code{cell.clusters} will be considered.}

\item{value}{String specifying if final resuls should be proportions ("percentage", default) or raw counts ("counts").}

\item{time.var}{Survival time variable.}

\item{event.var}{Variable with event censoring.}

\item{xerror}{Numeric value with cross-validation error for pruning the tree. If missing, the minimal error will be automatically selected.}
  
\item{rpart.control}{List with parameters for controling \code{rpart}. See \code{?rpart::rpart.control} for more information and defaults values (caution because, in \code{FlowCT} there are some changes: \code{cp = 0} and \code{minsplit = 10}).}

\item{return.data}{Original, pruned and used metadata should be returned?. Default = \code{TRUE}.}
}
\description{
This function performs a recursive partitioning tree using the percentage (or raw counts) of identified cell populations. Depending if \code{time.var} is present or absent, final tree were a cassification or regression tree, respectively.
}
\examples{
\dontrun{
tr <- rpart.tree(fcs.SCE = fcs, cell.clusters = "clusters_named", 
   time.var = "PFS", event.var = "PFS_c")
}
}
\keyword{CART}
\keyword{classification}
\keyword{partioning}
\keyword{recursive}
\keyword{regression}
\keyword{tree}
