% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sc.heatmap}
\alias{sc.heatmap}
\title{This function draws a heatmap with single-cell fluorescence values.}
\usage{
sc.heatmap(fcs.SE, assay.i = "normalized", markers.to.use = "all",
  not.metadata = c("filename", "cell_id", "sample_id"),
  clustering.method = "average", subsampling = 100, color = NULL)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{assay.i}{Name of matrix stored in the \code{FCS.SE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{markers.to.use}{Vector with markers to use. Default = \code{"all"}.}

\item{not.metadata}{Vector with variable names (from \code{colData(fcs.SE)}) for not including in the heatmap annotation. Default = \code{c("filename", "cell_id", "sample_id")}.}

\item{clustering.method}{Clustering method for rows and columns clustering within the heatmap. Possible values are "average" (default), "ward.D", "ward.D2", "single", "complete", "mcquitty", "median" or "centroid".}

\item{subsampling}{Numeric value indicating how many events use to draw heatmap and speed up plotting. Default = \code{100}.}

\item{color}{Color vector for coloring expression values within the heatmap. Default = \code{NULL} (i.e., scale "YlGnBu" from \code{RColorBrewer}).}
}
\description{
This function draws a heatmap with single-cell fluorescence values.
}
\examples{
\dontrun{
sc.heatmap(fcs_se100)
}
}
\keyword{expression}
\keyword{heatmap}
\keyword{single-cell}
