% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.exprs.R
\name{scale.exprs}
\alias{scale.exprs}
\title{Scale expression}
\usage{
\method{scale}{exprs}(
  data,
  assay.i = "normalized",
  include.FCS.SCE = F,
  scaled.matrix.name = "scaled"
)
}
\arguments{
\item{data}{A \code{fcs.SCE} object generated through \code{\link[FlowCT.v2:fcs.SCE]{FlowCT.v2::fcs.SCE()}} or a expression table with events in rows and markers in columns.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{include.FCS.SCE}{Logical indicating if new scaled matrix must to be stored within the \code{fcs.SCE} object.}

\item{scaled.matrix.name}{New scaled matrix name (it will stored within the \code{fcs.SCE} object). Default = \code{"scaled"}.}
}
\description{
It scales a matrix expression and fit all values to a range from 0 to 1.
}
\examples{
\dontrun{
# option 1: save scaled matrix within the fcs.SCE object (if provided)
fcs <- scale.exprs(fcs.SCE = fcs)

# option 2: scale an external data.frame
markers_expresion_scaled <- scale.exprs(fcs.SCE = markers_expresion)
}
}
\keyword{scale}
