% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMetadata.fromFCS.R
\name{scMetadata.fromFCS}
\alias{scMetadata.fromFCS}
\title{Create a single-event metadata}
\usage{
scMetadata.fromFCS(flowset, metadata, add.exprs = T)
}
\arguments{
\item{flowset}{A \code{flowset} object generated following the \href{https://bioconductor.org/packages/devel/bioc/vignettes/flowCore/inst/doc/HowTo-flowCore.pdf}{\code{flowCore} tutorial}.}

\item{metadata}{Metadata information for each read FCS file.}

\item{add.exprs}{Logical indicating if expression data must be added to each cell's metadata. Default = \code{TRUE}.}
}
\description{
It creates, from an initial metadata provided by the user, a global metadata for each event in a \href{https://www.rdocumentation.org/packages/flowCore/versions/1.38.2/topics/flowSet-class}{\code{flowset} object}.
}
\examples{
\dontrun{
 sc_metadata <- scMetadata.fromFCS(fcs, metadata, add.exprs = T)
}
}
\keyword{flowset}
\keyword{metadata}
\keyword{single-cell}
