% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sub.samples.idx}
\alias{sub.samples.idx}
\title{sub.samples.idx}
\usage{
sub.samples.idx(data, colname_samples, samples_names, subsampling = 1000,
  set.seed = 333, verbose = TRUE)
}
\arguments{
\item{data}{A data.frame with the marker expression values (and metadata information binded) for each cell.}

\item{colname_samples}{Colname of the column with samples indentifiers.}

\item{samples_names}{A vector with sample identifiers.}

\item{subsampling}{Number of events to keep. Default = 1000}

\item{set.seed}{\code{\link[base:set.seed]{set.seed()}} for random picking. Default = 333}

\item{verbose}{Logical indicating if display process output. Default = \code{TRUE}}
}
\description{
This function generates an index taking (in a random manner) a number of given events by sample.
}
\examples{
\dontrun{
sub_idx_som <- sub.samples.idx(data = mdsc_som, colname_samples = "sample_id",
    samples_names = md$sample_id, subsampling = 1000)
sel_expr <- mdsc_som[sub_idx_som,]}
}
\keyword{subsampling}
