% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.cutoff.R
\name{surv.plot}
\alias{surv.plot}
\title{Survival plots according cell cutoffs}
\usage{
surv.plot(pop.cutoff.obj, time.var, event.var, palette = "jco", variables, curve.type = "survival")
}
\arguments{
\item{pop.cutoff.obj}{An object generated through \code{\link[FlowCT.v2:pop.cutoff]{FlowCT.v2::pop.cutoff()}}.}

\item{time.var}{Survival time variable.}

\item{event.var}{Variable with event censoring.}

\item{palette}{Colors vector. Default = \code{"jco"}.}

\item{variables}{Vector with variables for calculating the cutoff. If nothing is detailed (\code{NULL}, default), all immune populations from \code{cell.clusters} will be considered.}

\item{curve.type}{Plotting curve methodology, i.e. "survival" (default) or "cumulative" (for progressions).}

}
\description{
Once immune cell cutoffs are calculated, this function draw the subsequent Kaplan-Meier curves for specified populations.
}
\examples{
\dontrun{
surv.plot(pop_cuts, time.var = "PFS", event.var = "PFS_c", curve.type = "survival")
}
}
\keyword{survival}
\keyword{cutoffs}
\keyword{Kaplan-Maier}
\keyword{curves}
