% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.tree.R
\name{surv.tree}
\alias{surv.tree}
\title{Kaplan-Meier from survival tree}
\usage{
surv.tree(
  rpart.tree,
  tree,
  time.var,
  event.var,
  return.data = F,
  curve.type = "survival"
)
}
\arguments{
\item{rpart.tree}{A object generated through \code{\link[FlowCT.v2:rpart.tree]{FlowCT.v2::rpart.tree()}}.}

\item{tree}{Specify the tree name, i.e. "original" or "pruned".}

\item{time.var}{Survival time variable.}

\item{event.var}{Variable with event censoring.}

\item{return.data}{Metadata with node leafs distribution should be returned. Default = \code{FALSE}.}

\item{curve.type}{Plotting curve methodology, i.e. "survival" (default) or "cumulative" (for progressions).}
}
\description{
Plot with a Kaplan-Meier curve with node leafs identified by recursive partitioning.
}
\examples{
\dontrun{
surv.tree(rpart.tree = tr, tree = "original", time.var = "PFS", event.var = "PFS_c")
}
}
\keyword{CART}
\keyword{Kaplan-Meier}
\keyword{classification}
\keyword{curves}
\keyword{cutoffs}
\keyword{partioning}
\keyword{recursive}
\keyword{regression}
\keyword{survival}
\keyword{tree}
