% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxregs.R
\name{unicox}
\alias{unicox}
\title{Univariable Cox Regression}
\usage{
unicox(
  data,
  covariates,
  pval.cutoff = 0.05,
  timevar,
  censvar,
  xlim = c(0, 6),
  log.scale = F,
  hide.nosig = F
)
}
\description{
Univariable Cox Regression
}
