% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.flowSet.SE.R
\name{as.flowSet.SE}
\alias{as.flowSet.SE}
\title{Transform a \code{fcs.SCE} object into a \code{flowset} object}
\usage{
as.flowSet.SE(fcs.SCE, assay.i)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT.v2:fcs.SCE]{FlowCT.v2::fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which obtain a \code{flowset} object.}
}
\description{
It tranforms a \code{fcs.SCE} object into a \href{https://www.rdocumentation.org/packages/flowCore/versions/1.38.2/topics/flowSet-class}{\code{flowset}} object.
}
\keyword{fcs.SCE}
\keyword{flowset}
\keyword{generation}
\keyword{to}
