% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as.flowSet.SE}
\alias{as.flowSet.SE}
\title{as.flowSet.SE}
\usage{
as.flowSet.SE(fcs.SE, assay.i)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{assay.i}{Name of matrix stored in the \code{FCS.SE} object from which obtain a \code{flowset} object.}
}
\description{
It tranforms a FCS.SE object into a (\href{https://www.rdocumentation.org/packages/flowCore/versions/1.38.2/topics/flowSet-class}{flowset}) object.
}
\examples{
\dontrun{
 sc_metadata <- scMetadata.fromFCS(fcs, metadata, add.exprs = T)
}
}
\keyword{FCS.SE}
\keyword{flowset}
\keyword{generation}
\keyword{to}
