% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{barplot.cell.pops}
\alias{barplot.cell.pops}
\title{barplot.cell.pops}
\usage{
\method{barplot}{cell.pops}(cell_clusters, metadata, colname_sampleID)
}
\arguments{
\item{cell_clusters}{A vector with clusters identified in a previous clustering method.}

\item{metadata}{A data.frame with additional information for cell proportions (it must include a column with the condition and named as "condition").}

\item{colname_sampleID}{Colname of the column containing sample identification.}
}
\description{
This function calculates cluster proportions for each identified cluster and draws it over a stacked barplot by condition.
}
\examples{
\dontrun{
props_final <- barplot.cell.pops(cell_clusters = final_metadata$final_cluster,
    metadata = final_metadata[,1:3], colname_sampleID = "sample_id")}
}
\keyword{barplot}
\keyword{proportions}
