% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.cell.pops.R
\name{barplot.cell.pops}
\alias{barplot.cell.pops}
\title{Stacked barplot with cell clusters}
\usage{
\method{barplot}{cell.pops}(
  fcs.SCE,
  assay.i = "normalized",
  cell.clusters,
  plot = T,
  count.by,
  facet.by = NULL,
  return.mode,
  colors = NULL
)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT:fcs.SCE]{FlowCT::fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{cell.clusters}{Name of column containing clusters identified through \code{\link[FlowCT:clustering.flow]{FlowCT::clustering.flow()}}.}

\item{plot}{Logical indicating whether plotting stacked barplot. Default = \code{TRUE}.}

\item{count.by}{Variable name (from \code{colData(fcs.SCE)}) for calculating proportions (or counts) and drawing the x-axis in the stacked bar plotting.}

\item{facet.by}{Variable name (from \code{colData(fcs.SCE)}) for splitting the stacked bar plotting. Default = \code{NULL}.}

\item{return.mode}{String for specifying if final resuls should be proportions ("percentage") or raw counts ("counts"). Default = \code{NULL}.}

\item{colors}{Vector with colors for plotting. Default = \code{NULL} (i.e., it will choose automatically a vector of colors according to \code{\link[FlowCT:div.colors]{FlowCT::div.colors()}}).}
}
\description{
This function calculates cluster proportions (or raw counts) for each identified cluster and plot them on a stacked barplot.
}
\examples{
\dontrun{
prop_table <- barplot.cell.pops(fcs.SCE = fcs, cell.clusters = "SOM_named", 
    count.by = "sample_id", facet.by = "condition", 
    return.mode = "percentage")
counts_table <- barplot.cell.pops(fcs.SCE = fcs, cell.clusters = "SOM_named", 
    count.by = "condition", return.mode = "counts")
}
}
\keyword{barplot}
\keyword{proportions}
