% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{boxplot.cell.clustering}
\alias{boxplot.cell.clustering}
\title{It draws a boxplot with cell clusters identified through \code{\link[FlowCT:fsom.clustering]{FlowCT::fsom.clustering()}}.}
\usage{
\method{boxplot}{cell.clustering}(fcs.SE, assay.i = "normalized",
  cell.clusters, condition.column = "condition",
  pvalue.cutoffs = list(cutpoints = c(0, 1e-04, 0.001, 0.01, 0.05, 1),
  symbols = c("xxxx", "***", "**", "*", "ns")),
  color.by = condition.column, geom.point = T, facet = F,
  facet.free.scale = "free_x", shape.by = NULL, y.limits = NULL,
  show.stats = T, return.stats = T, plot.only.sig = c(F, 0.05))
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{assay.i}{Name of matrix stored in the \code{FCS.SE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{cell.clusters}{A vector with clusters identified through \code{\link[FlowCT:fsom.clustering]{FlowCT::fsom.clustering()}} (and, normaly, later renamed).}

\item{condition.column}{Column name from the \code{colData(fcs.SE)} object which contains condition information.}

\item{pvalue.cutoffs}{List of P-value cutoffs and their symbols for indicante significances within the plot. Default = \code{list(cutpoints = c(0, 0.0001, 0.001, 0.01, 0.05, 1), symbols = c("xxxx", "***", "**", "*", "ns"))}.}

\item{color.by}{Variable name (from \code{colData(fcs.SE)}) for lines coloring.}

\item{geom.point}{Logical indicating if adding points to boxplot. Default = \code{TRUE}.}

\item{facet}{Logical indicating if splitting boxplots by cell clusters. Default = \code{FALSE}.}

\item{facet.free.scale}{If \code{facet = TRUE}, string indicating how scales would be shared across all facets. Possible values: "free_x" (default), "free_y" and "free".}

\item{shape.by}{Variable name (from \code{colData(fcs.SE)}) for dot shaping. Default = \code{NULL}.}

\item{y.limits}{Numeric vector with limits for y-axis (minimum, maximum). Default = \code{NULL}.}

\item{show.stats}{Significances should be added to boxplots? Default = \code{TRUE}.}

\item{return.stats}{Logical indicating if calculated statistics should be returned in a new variable. Default = \code{TRUE}.}

\item{plot.only.sig}{Vector indicating if only significant cell clusters should be displayed (logical element) and the P-value cutoff for selecting those ones (numerical element). Default = \code{c(F, 0.05)}.}
}
\description{
It draws a boxplot with cell clusters identified through \code{\link[FlowCT:fsom.clustering]{FlowCT::fsom.clustering()}}.
}
\examples{
\dontrun{
# option 1: show all cell clusters and return statistics
bx_sig <- boxplot.cell.clustering(fcs.SE = fcs_se, cell.clusters = fcs_se$SOM_named, facet = T, 
    return.stats = T, 
    facet.free.scale = "free")

# option 2: show only those significant cell clusters
boxplot.cell.clustering(fcs.SE = fcs_se, cell.clusters = fcs_se$SOM_named, return.stats = F, 
    plot.only.sig = c(T, 0.1))
}
}
\keyword{boxplot}
\keyword{cell}
\keyword{clusters}
\keyword{differential}
\keyword{distributions}
