% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cell.count.bx}
\alias{cell.count.bx}
\title{cell.count.bx}
\usage{
cell.count.bx(data, counts_by = "sample_id", metadata, labels = NULL)
}
\arguments{
\item{data}{A data.frame with the metadata for each cell.}

\item{counts_by}{Condition to perform the counting. Default = "sample_id"}

\item{metadata}{A data.frame with additional information for cell proportions.}

\item{labels}{Labels to add to the points on the boxplot, it must be taken from \code{metadata}. Default = \code{NULL}}
}
\description{
This function draws a boxplot according to the cell count for a specified condition.
}
\examples{
\dontrun{cell.count.bx(metadata_sc, counts_by = "sample_id", metadata = md)}
}
\keyword{boxplot}
\keyword{cell}
\keyword{count}
