% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.count.bx.R
\name{cell.count.bx}
\alias{cell.count.bx}
\title{Boxplot according according metadata}
\usage{
cell.count.bx(
  fcs.SCE,
  assay.i = "normalized",
  x.axis,
  color.by = x.axis,
  label.by = NULL,
  limits = NULL,
  colors = NULL
)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT:fcs.SCE]{FlowCT::fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{x.axis}{Variable name (from \code{colData(fcs.SCE)}) for drawing the x-axis in plot.}

\item{color.by}{Variable name (from \code{colData(fcs.SCE)}) for coloring the plot. Default = \code{x.axis} (i.e., the same specfied in \code{x.axis}).}

\item{label.by}{Variable from (from \code{colData(fcs.SCE)}) for labeling. Default = \code{NULL}.}

\item{limits}{Numeric vector with limits for plotting (minimum, maximum). Default = \code{NULL} (i.e., automatically calculated from data).}

\item{colors}{Vector with colors for plotting. Default = \code{NULL} (i.e., it will choose automatically a vector of colors according to \code{\link[FlowCT:div.colors]{FlowCT::div.colors()}}).}
}
\description{
This function draws a boxplot according to the cell count for a specified condition.
}
\examples{
\dontrun{cell.count.bx(fcsL, assay.i = "normalized", x.axis = "condition")}
}
\keyword{boxplot}
\keyword{cell}
\keyword{count}
