% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{circ.tree}
\alias{circ.tree}
\title{circ.tree}
\usage{
circ.tree(expr, expr_saturated, cell_clusters, dist_method = "euclidean",
  hclust_method = "average", open_angle = 100, dendro_labels = FALSE,
  nodes = NULL, set.seed = 333)
}
\arguments{
\item{expr}{A data.frame with marker expression values for each cell}

\item{expr_saturated}{A data.frame with marker expression values for each cell where those expression values greater or smaller that 1 and 0, respectively, must be coherced to 1 and 0.}

\item{cell_clusters}{A vector with clusters identified in a previous clustering method.}

\item{dist_method}{The distance measure to be used. Possible values are "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".}

\item{hclust_method}{The agglomeration method to be used. Possible values are "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median" or "centroid".}

\item{open_angle}{Open angle for circular layout. Default = 100}

\item{dendro_labels}{Logical whether adding labels to dendrogram. Default = FALSE}

\item{nodes}{Those nodes limiting different color areas. Default = NULL}

\item{set.seed}{\code{\link[base:set.seed]{set.seed()}}, for selecting colors for color areas. Default = 333}
}
\description{
This function plots a circular dendrogram and a heatmap according expression clusters previously calculated. Every leaf has a different point size regarding the frequency for each cluster.
For additional information go to \href{https://guangchuangyu.github.io/software/ggtree/documentation/}{\code{\link{ggtree}}} package.
}
\examples{
\dontrun{
circ.tree(expr = matrix_clusters, expr_saturated = matrix_clusters01,
    cell_clusters = cell_clustering, dendro_labels = TRUE)
circ.tree(expr = matrix_clusters, expr_saturated = matrix_clusters01,
    cell_clusters = cell_clustering, nodes = c(29,33,35))}
}
\keyword{circular}
\keyword{dendrogram}
\keyword{node}
\keyword{tree}
