% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.plot.conditions.R
\name{corplot.conditions}
\alias{corplot.conditions}
\title{cor.plot.conditions}
\usage{
corplot.conditions(fcs.SCE, assay.i = "normalized", cell.clusters,
  condition.column, return.stats = F)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT:fcs.SCE]{FlowCT::fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{cell.clusters}{A vector with clusters identified through \code{\link[FlowCT:fsom.clustering]{FlowCT::fsom.clustering()}} (and, normaly, later renamed).}

\item{condition.column}{Column name from the \code{colData(fcs.SCE)} object which contains condition information.}

\item{return.stats}{Logical indicating if calculated statistics should be returned in a new variable. Default = \code{FALSE}.}
}
\description{
It draws a correlation plot (between two specified conditions within a \code{fcs.SCE} object.
}
\examples{
\dontrun{
corplot.conditions(fcs.SCE = fcs, cell.clusters = fcs$SOM_named, condition.column = "condition")
}
}
\keyword{corr}
\keyword{correlation}
\keyword{plot}
