% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_and_clustering.R
\name{dim.reduction}
\alias{dim.reduction}
\title{dim.reduction}
\usage{
\method{dim}{reduction}(data, assay.i = "normalized", metadata = NULL,
  markers.to.use = "all", dr.method = c("all", "PCA", "tSNE", "UMAP"),
  num.threads = NULL, perplexity.tsne = 100, n.neighbors.umap = 50)
}
\arguments{
\item{data}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}} or a expression table with events in rows and markers in columns.}

\item{assay.i}{Name of matrix stored in the \code{FCS.SE} object from which calculate the DR (this option is useless if input is not a \code{FCS.SE} object). Default = \code{"normalized"}.}

\item{metadata}{If \code{data} is an expression matrix, an additional metadata table with information for each FCS file is necessary.}

\item{markers.to.use}{Markers to take account in the DR calculus. Default = \code{"all"}.}

\item{dr.method}{DR method to calculate. Possible values are "PCA", "tSNE" or "UMAP", individually, or "all".}

\item{num.threads}{Number of threads for DR calculus. If \code{NULL} (default), all cores available minus one will be used.}

\item{perplexity.tsne}{Value for perplexity parameter in tSNE calculation (\href{https://distill.pub/2016/misread-tsne/}{more information}). Default = \code{100}.}

\item{n.neighbors.umap}{Value for neighbors parameter in UMAP calculation (\href{https://www.math.upenn.edu/~jhansen/2018/05/04/UMAP/}{more information}). Default = \code{50}.}
}
\value{
The final output is an object of type \code{list} with two elements \enumerate{
          \item \code{dr}: A data.frame combining metadata and expression values, together with dimensional reduction data.
          \item \code{dr_melted}: A data.frame combining metadata and expression values, together with dimensional reduction data but transformed to long format by marker for later plotting.}
}
\description{
It calculates a dimensional reduction (DR) from a \code{FCS.SE} object (or an expression table). Three different DR methods are available: Principal Component Analsis (PCA), t-Distributed Stochastic Neighbor Embedding (t-SNE) and  Uniform Manifold Approximation and Projection (UMAP).
}
\examples{
\dontrun{
dr_tsne <- dim.reduction(fcs_se, dr.method = "tSNE", 
    markers.to.use = c("CD8", "CD27", "CCR4", "CD45RA", "CD4"))
dr_umap <- dim.reduction(fcs_se, dr.method = "UMAP", n.neighbors.umap = 10)
dr_all <- dim.reduction(fcs_se, dr.method = "all")
}
}
\keyword{PCA}
\keyword{UMAP}
\keyword{dimensional}
\keyword{reduction}
\keyword{tSNE}
