% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.reduction.R
\name{dim.reduction}
\alias{dim.reduction}
\title{Calculate dimensional reductions}
\usage{
\method{dim}{reduction}(
  data,
  assay.i = "normalized",
  markers.to.use = "all",
  dr.method,
  num.threads = NULL,
  perplexity.tsne = 100,
  n.neighbors.umap = 50
)
}
\arguments{
\item{data}{A \code{fcs.SCE} object generated through \code{\link[FlowCT:fcs.SCE]{FlowCT::fcs.SCE()}} or a expression table with events in rows and markers in columns.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate the DR (this option is useless if input is not a \code{fcs.SCE} object). Default = \code{"normalized"}.}

\item{markers.to.use}{Markers to take account in the DR calculus. Default = \code{"all"}.}

\item{dr.method}{DR method to calculate. Possible values are "PCA", "tSNE", "DENSNE", "UMAP" and/or "DENSMAP".}

\item{num.threads}{Number of threads for DR calculus. If \code{NULL} (default), all cores available minus one will be used. If this option is enable, reproducibility could be comprised.}

\item{perplexity.tsne}{Value for perplexity parameter in tSNE calculation (\href{https://distill.pub/2016/misread-tsne/}{more information}). Default = \code{100}.}

\item{n.neighbors.umap}{Value for neighbors parameter in UMAP calculation (\href{https://www.math.upenn.edu/~jhansen/2018/05/04/UMAP/}{more information}). Default = \code{50}.}
}
\description{
It calculates a dimensional reduction (DR) from a \code{fcs.SCE} object (or an expression table). Three different DR methods are available: Principal Component Analsis (PCA), t-Distributed Stochastic Neighbor Embedding (t-SNE) and  Uniform Manifold Approximation and Projection (UMAP).
}
\examples{
\dontrun{
fcs <- dim.reduction(fcs, dr.method = "tSNE", 
   markers.to.use = c("CD8", "CD27", "CCR4", "CD45RA", "CD4"))
fcs <- dim.reduction(fcs, dr.method = "UMAP", n.neighbors.umap = 10)
fcs <- dim.reduction(fcs, dr.method = c("pca", "Umap"))
}
}
\keyword{PCA}
\keyword{UMAP}
\keyword{dimensional}
\keyword{reduction}
\keyword{tSNE}
