% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_and_clustering.R
\name{dim.reduction}
\alias{dim.reduction}
\title{dim.reduction}
\usage{
\method{dim}{reduction}(expr_data, metadata, reduction_method = c("all",
  "PCA", "tSNE", "UMAP"), num_threads = 4, tsne_perplexity = 100,
  n_neighbors_umap = 50, set.seed = 333)
}
\arguments{
\item{expr_data}{A data.frame with marker expression values for each sample, patient, cell... (ie. cols are flow markers and rows are samples/cells).}

\item{metadata}{A data.frame with additional information for \code{expr_data}.}

\item{reduction_method}{Dimensional reduction method. Possible values are "PCA", "tSNE", "UMAP" or "all".}

\item{num_threads}{Number of threads. Default = 4}

\item{tsne_perplexity}{Value for perplexity parameter in tSNE calculation (\href{https://distill.pub/2016/misread-tsne/}{more information}). Default = 100}

\item{n_neighbors_umap}{Value for neighbors parameter in UMAP calculation (\href{https://www.math.upenn.edu/~jhansen/2018/05/04/UMAP/}{more information}). Default = 50}

\item{set.seed}{\code{\link[base:set.seed]{set.seed()}} for dimensional reduction. Default = 333}
}
\value{
The final output is an object of type list with two elements \enumerate{
          \item \code{dr}: A data.frame with input table (cols = flow markers) together with dimensional reduction data,
          \item \code{dr_melted}: A data.frame with input table together with dimensional reduction data but collapsed to long format by marker.}
}
\description{
This function calculates cluster proportions for each identified cluster and draws it over a stacked barplot by condition.
}
\examples{
\dontrun{
dr <- dim.reduction(sel_expr, metadata = sel_exprL, reduction_method = "all",
    set.seed = 1234)
    }
}
\keyword{PCA}
\keyword{UMAP}
\keyword{dimensional}
\keyword{reduction}
\keyword{tSNE}
