% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.plotting.R
\name{dr.plotting}
\alias{dr.plotting}
\title{dr.plotting}
\usage{
dr.plotting(data, assay.i = "normalized", plot.dr, dims = c(1, 2),
  color.by = "expression", shape.by = NULL, facet.by = NULL, omit.markers = NULL,
  title = "", label.by = NULL, size = 2, raster = c(FALSE, 1000),
  return.df = FALSE, colors = NULL)
}
\arguments{
\item{data}{A object with DR generated with \code{\link[FlowCT.v2:dim.reduction]{FlowCT.v2::dim.reduction()}} or a \code{data.frame} with DR, expression and metadata information (like the first element list of the object generated with \code{\link[FlowCT.v2:dim.reduction]{FlowCT.v2::dim.reduction()}}).}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{plot.dr}{String indicating the desired DR to plot (this indicated DR should be prevoulsy calculated to being plotted).}

\item{color.by}{Variable from (from \code{colData(fcs.SCE)}) for dots coloring. If \code{color.by = "expression"} (default), plot will be automatically splitted for each marker.}

\item{shape.by}{Variable from (from \code{colData(fcs.SCE)}) for dots shaping. Default = \code{NULL}.}

\item{facet.by}{Variable from (from \code{colData(fcs.SCE)}) for plot spliting. Default = \code{NULL}.}

\item{omit.markers}{Vector with markers to omit when plotting with \code{color.by = "expression"}. Default = \code{NULL}.}

\item{title}{Title to add to the plot.}

\item{label.by}{Variable from (from \code{colData(fcs.SCE)}) for dots labeling. Default = \code{NULL}.}

\item{size}{Point size. Default = \code{1}.}

\item{raster}{Vector indicating if image should be rasterized (logical element) and the number of pixels to consider (numerical element). It is based on \href{https://github.com/exaexa/scattermore}{\code{scattermore} package}. Default = \code{c(FALSE, 1000)}.}

\item{return.df}{Logical indicating if built \code{data.frame} with DR information and metadata must be returned. Default = \code{FALSE}.}

\item{n.dims}{Vector indicating the two DR components to plot. Default = \code{c(1,2)}.}

\item{colots}{Vector with colors for plotting. Default = \code{NULL} (i.e., it will choose automatically a vector of colors according to \code{\link[FlowCT.v2:div.colors]{FlowCT.v2::div.colors()}}).}
}
\description{
This function plots the indicated dimensional reduction (DR) from a previously calculated \code{\link[FlowCT.v2:dim.reduction]{FlowCT.v2::dim.reduction()}} object.
}
\examples{
\dontrun{
dr.plotting(fcs, plot.dr = "tSNE", color.by = "condition")
dr.plotting(fcs, plot.dr = "UMAP", color.by = "patient_id")
dr <- dr.plotting(fcs, plot.dr = "PCA", color.by = "SOM", facet.by = "condition", return.df = T)
}
}
\keyword{PCA}
\keyword{UMAP}
\keyword{dimensional}
\keyword{plotting}
\keyword{reduction}
\keyword{tSNE}
