% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dr.plotting}
\alias{dr.plotting}
\title{dr.plotting}
\usage{
dr.plotting(data, dr_calculated, color_by = "expression",
  facet_by = NULL, labels = NULL, size = 0.5, output_name = "",
  output_type = NULL)
}
\arguments{
\item{data}{A data.frame generated with \code{\link[FlowCT:dim.reduction]{FlowCT::dim.reduction()}}.}

\item{dr_calculated}{Type of dimensional reduction for plotting. Possible values are "PCA", "tSNE" or "UMAP".}

\item{color_by}{Variable for coloring the plot. Default = "expression"}

\item{facet_by}{Variable for faceting the plot. Default = \code{NULL}}

\item{labels}{Variable for labeling points. Default = \code{NULL}}

\item{size}{Point size. Default = 0.5}

\item{output_name}{Preffix to add to the final saved file. Default = ""}

\item{output_type}{Format for saving plot. Possible values are \code{NULL} (Default, i.e. the plot is shown in the \emph{Plots panel}), "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf".}
}
\description{
This function plots a selected dimensional reduction from those previously calculated with FlowCT::dim.reduction().
}
\examples{
\dontrun{
dr.plotting(dr$dr_melted, dr_calculated = "PCA", color_by = "patient",
    output_type = NULL)
dr.plotting(dr$dr_melted, dr_calculated = "tSNE", color_by = "patient",
    facet_by = "condition", output_type = "png")}
}
\keyword{PCA}
\keyword{UMAP}
\keyword{dimensional}
\keyword{reduction}
\keyword{tSNE}
