% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{export.metaFCS}
\alias{export.metaFCS}
\title{export.metaFCS}
\usage{
export.metaFCS(fcs.SE, dr.object, output.name)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{dr.object}{Object created with \code{\link[FlowCT:dim.reduction]{FlowCT::dim.reduction()}} function or a table combining DR and experimental metadata information.}

\item{output.name}{Name for generated FCS file. Important, add the final extension ".fcs".}
}
\description{
It creates a FCS file containing all analysis incorpored to colData(fcs.SE) as well as the dimensional reduction coordinates calculated.
}
\examples{
\dontrun{
export.metaFCS(fcs.SE = fcs_se, dr.object = dr, output.name = "final_file.fcs")
}
}
\keyword{FCS}
\keyword{exporting}
\keyword{final}
\keyword{generation}
