% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.norm.R
\name{gauss.norm}
\alias{gauss.norm}
\title{gauss.norm}
\usage{
gauss.norm(fcs.SCE, assay.i = "transformed", marker.to.norm,
  norm.matrix.name = "normalized")
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT]{fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"transformed"}.}

\item{marker.to.norm}{Marker name(s) to normalize (based on visual inspection with \code{\link[FlowCT]{multidensity()}}).}

\item{norm.matrix.name}{New normalized matrix name (it will stored within the \code{fcs.SCE} object). Default = \code{"normalized"}.}
}
\description{
It performs a Gaussian normalization based on \code{\link[flowStats]{gaussNorm()}} function.
}
\examples{
\dontrun{
 fcs <- gauss.norm(fcs.SCE = fcs, marker.to.norm = c("CD62L", "CCR4", "SSC_A"))
}
}
\keyword{Gaussian}
\keyword{alignment}
\keyword{marker}
\keyword{normalization}
