% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lims.FCS}
\alias{lims.FCS}
\title{lims.FCS}
\usage{
lims.FCS(fcs)
}
\arguments{
\item{fcs}{A \code{flowSet} object read with \code{\link[flowCore:read.flowSet]{flowCore::read.flowSet()}}.}
}
\description{
This function estimates the minimum and maximum values for an object with multiple FCS files. Useful for plotting limits.
}
\examples{
\dontrun{lims.FCS(fcs_raw)}
}
