% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.normalization.R
\name{marker.normalization}
\alias{marker.normalization}
\title{Normalization for markers' expression}
\usage{
marker.normalization(
  fcs.SCE,
  assay.i = "transformed",
  method,
  marker,
  new.matrix.name = "normalized"
)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT.v2:fcs.SCE]{FlowCT.v2::fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"transformed"}.}

\item{method}{Methodology to perform each marker normalization. Possible values are "gauss" or "warp" (from \code{flowStats} package).}

\item{marker}{Marker name(s) to normalize (based on visual inspection with \code{\link[FlowCT.v2:multidensity]{FlowCT.v2::multidensity()}}).}

\item{new.matrix.name}{New normalized matrix name (it will stored within the \code{fcs.SCE} object). Default = \code{"normalized"}.}
}
\description{
It normalizes expression abnormalities within selected markers.
}
\examples{
\dontrun{
 fcs_gauss <- normalization.flw(fcs.SCE = fcs, marker = c("CD62L", "CCR4", "SSC_A"), method = "gauss")
 fcs_warp <- normalization.flw(fcs.SCE = fcs, marker = c("CD62L", "CCR4", "SSC_A"), method = "warp")
}
}
\keyword{alignment}
\keyword{gaussian}
\keyword{marker}
\keyword{normalization}
\keyword{warp}
