% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{marker.normalization}
\alias{marker.normalization}
\title{Marker normalization}
\usage{
  marker.normalization(fcs.SCE, assay.i = "transformed", method, marker,
                       new.matrix.name = "normalized")
}
\arguments{
  \item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

  \item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"transformed"}.}

  \item{marker}{Marker name(s) to normalize (based on visual inspection with \code{\link[FlowCT:multidensity]{FlowCT::multidensity()}}).}

  \item{method}{Methodology to perform each marker normalization. Possible values are "gauss" or "warp" (from \code{flowStats} package).}

  \item{new.matrix.name}{New normalized matrix name (it will stored within the \code{FCS.SE} object). Default = \code{"normalized"}.}
}
\description{
  It normalizes expression abnormalities within selected markers.
}
\examples{
  \dontrun{
    fcs_gauss <- normalization.flw(fcs.SCE = fcs, marker = c("CD62L", "CCR4", "SSC_A"), method = "gauss")
    fcs_warp <- normalization.flw(fcs.SCE = fcs, marker = c("CD62L", "CCR4", "SSC_A"), method = "warp")
  }
}
\keyword{alignment}
\keyword{marker}
\keyword{normalization}
\keyword{gaussian}
\keyword{warp}
