% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.dr.R
\name{median.dr}
\alias{median.dr}
\title{Calculate a DR with median values}
\usage{
\method{median}{dr}(
  fcs.SCE,
  assay.i = "normalized",
  markers.to.use = "all",
  dr.method = "PCA",
  num.k = NULL,
  perplexity.tsne = 15,
  n.neighbors.umap = 50,
  color.by = "filename",
  shape.by = NULL,
  label.by = NULL,
  size = 3,
  return.DR = F
)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT:fcs.SCE]{FlowCT::fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{markers.to.use}{Vector with markers to use. By default (\code{"all"}), all markers will be used.}

\item{dr.method}{Possible values are "PCA" (default), "tSNE", "UMAP", "DENSMAP" or "DENSNE". Take into account the number of samples for this DR, for few samples t-SNE and UMAP are useless. If `pca.loadings`, the plotting result will show the weight of each marker for each PC (the first two components).}

\item{num.k}{Number of clusters if \code{dr.method = kmeans}. If \code{NULL} (default), a classical plot is drawn, but if a number is indicated, K-means clustering is performed and displayed as ellipse-rounded groups.}

\item{perplexity.tsne}{Value for perplexity parameter in tSNE calculation (\href{https://distill.pub/2016/misread-tsne/}{more information}). For little number of samples this DR (t-SNE) cannot be calculated because preplexity restrictions. Default = \code{15}.}

\item{n.neighbors.umap}{Value for neighbors parameter in UMAP calculation (\href{https://www.math.upenn.edu/~jhansen/2018/05/04/UMAP/}{more information}). Default = \code{50}.}

\item{color.by}{Variable name (from \code{colData(fcs.SCE)}) for dots coloring. Default = \code{"filename"}.}

\item{shape.by}{Variable name (from \code{colData(fcs.SCE)}) for dots shaping. Default = \code{"NULL"}.}

\item{label.by}{Variable name (from \code{colData(fcs.SCE)}) for labelling. Default = \code{NULL}.}

\item{size}{Point size. Default = \code{3}.}

\item{return.DR}{Final data combining metadata, K-means and DR info should be returned?. Default = \code{FALSE}.}
}
\description{
This function performs a dimensional reduction over median expressino values for each file within the \code{fcs.SCE} object and, if user specifies, calculated K-means clustering.
}
\examples{
\dontrun{
median.dr(fcs, color.by = "condition")
}
}
\keyword{dimensional}
\keyword{expression}
\keyword{kmeans}
\keyword{median}
\keyword{reduction}
\keyword{values}
