% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.heatmap.R
\name{median.heatmap}
\alias{median.heatmap}
\title{median.heatmap}
\usage{
\method{median}{heatmap}(fcs.SCE, assay.i = "normalized",
  cell.clusters = NULL, markers.to.use = "all",
  not.metadata = "filename", colors = NULL)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT.v2:fcs.SCE]{FlowCT.v2::fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{cell.clusters}{A vector with clusters identified through \code{\link[FlowCT.v2:fsom.clustering]{FlowCT.v2::fsom.clustering()}} (and, normaly, later renamed). Default = \code{NULL} (i.e., median values will be calculated for each FCS file).}

\item{markers.to.use}{Vector with markers to use. Default = \code{"all"}.}

\item{not.metadata}{Vector with variable names (from \code{colData(fcs.SCE)}) for not including in the heatmap annotation. Default = \code{"filename"}.}

\item{colors}{Vector with colors for plotting (if provided, it must be as long as the number of unique elements in the longer metadata field). Default = \code{NULL} (i.e., it will choose automatically a vector of colors according to \code{\link[FlowCT.v2:div.colors]{FlowCT.v2::div.colors()}}).}
}
\description{
This function draws a heatmap with median values for each FCS file or for identified cluster with \code{\link[FlowCT.v2:fsom.clustering]{FlowCT.v2::fsom.clustering()}}
}
\examples{
\dontrun{
# option 1: general heatmap (by FCS file)
median.heatmap(fcs, not.metadata = c("sample_id", "file_name"))

# option 2: heatmap by SOM-identified clusters
median.heatmap(fcs.SCE = fcs, assay.i = "normalized", cell.clusters = fcs$SOM)
}
}
\keyword{cell}
\keyword{cluster}
\keyword{expression}
\keyword{heatmap}
\keyword{median}
\keyword{percentages}
\keyword{values}
