% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.values.R
\name{median.values}
\alias{median.values}
\title{It calculates median values according a specfied variable, normaly "filename".}
\usage{
\method{median}{values}(fcs.SCE, assay.i = "raw", var = "filename")
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT]{fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate medians. Default = \code{"raw"}.}

\item{var}{Variable for grouping and calculating medians. Default = \code{"filename"} (i.e., names of FCS files).}
}
\description{
It calculates median values according a specfied variable, normaly "filename".
}
\examples{
\dontrun{
mfis_FCS <- median.values(fcs.SCE = fcs)
med_SOM_clust <- median.values(fcs.SCE = fcs, assay.i = "normalized", var = "SOM_named")
}
}
\keyword{MFI}
\keyword{fluorescence}
\keyword{intensity}
\keyword{median}
\keyword{values}
