% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidensity.R
\name{multidensity}
\alias{multidensity}
\title{multidensity}
\usage{
multidensity(fcs.SCE, assay.i, show.markers = "all", color.by = NULL,
  subsampling = NULL, interactive = F, ridgeline.lim = 15, colors = NULL)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT.v2:fcs.SCE]{FlowCT.v2::fcs.SCE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"normalized"}.}

\item{show.markers}{Vector with markers to plot. Default = \code{"all"}.}

\item{color.by}{Variable name (from \code{colData(fcs.SCE)}) for lines coloring.}

\item{subsampling}{Numeric value indicating how many events use to calculate density lines and speed up plotting. Default = \code{NULL}.}

\item{interactive}{Logical indicating if the user can interact with the (only overlapping-lines) plot. Default = \code{FALSE}.}

\item{ridgeline.lim}{Numeric value specifying the limit for shifting from ridgeline-mode to overlapping-lines plot. Default = \code{15}.}

\item{colors}{Vector with colors for plotting (only available for no-ridgeline plot). Default = \code{NULL} (i.e., it will choose automatically a vector of colors according to \code{\link[FlowCT.v2:div.colors]{FlowCT.v2::div.colors()}}).}
}
\description{
This function draws multidensity plot with all FCS files included in a \code{fcs.SCE} object. If there are more files than limit specified in \code{ridgeline.lim}, instead of plottting density in a ridge-way all density lines will be overlapped.
}
\examples{
\dontrun{
multidensity(fcs.SCE = fcs, assay.i = "normalized", subsampling = 1000)
multidensity(fcs, assay.i = 2, color.by = "file_name", ridgeline.lim = 0, show.markers = c("CD62L", "CD4"), interactive = T)
}
}
\keyword{alignment}
\keyword{density}
\keyword{marker}
\keyword{normalization}
