% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{normalization.flw}
\alias{normalization.flw}
\title{normalization.flw}
\usage{
normalization.flw(fcs.SE, marker.to.norm)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{assay.i}{Name of matrix stored in the \code{fcs.SCE} object from which calculate correlation. Default = \code{"transformed"}.}

\item{marker.to.norm}{Marker name(s) to normalize (based on visual inspection with \code{\link[FlowCT:multidensity]{FlowCT::multidensity()}}).}

\item{norm.method}{Methodology to perform each marker normalization. Possible values are "gauss" or "warp" (from \code{flowStats} package) or "harmony" (from \code{harmony} package for single-cell).}

\item{var.to.use}{Variable from \code{colData(fcs.SCE)} for calculating the normalization (only if \code{norm.method = "harmony"}).}

\item{norm.matrix.name}{New normalized matrix name (it will stored within the \code{FCS.SE} object). Default = \code{"normalized"}.}
}
\description{
It normalizes expression abnormalities within selected markers.
}
\examples{
\dontrun{
 fcs_gauss <- normalization.flw(fcs.SCE = fcs, marker.to.norm = c("CD62L", "CCR4", "SSC_A"), norm.method = "gauss")
 fcs_warp <- normalization.flw(fcs.SCE = fcs, marker.to.norm = c("CD62L", "CCR4", "SSC_A"), norm.method = "warp")
 fcs_harmony <- normalization.flw(fcs.SCE = fcs, norm.method = "harmony", var.to.use = "condition")
}
}
\keyword{alignment}
\keyword{marker}
\keyword{normalization}
\keyword{gaussian}
\keyword{warp}
\keyword{harmony}
