% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.cutoff.R
\name{pop.cutoff}
\alias{pop.cutoff}
\title{Immune cutoffs}
\usage{
pop.cutoff(
  fcs.SCE,
  cell.clusters,
  value = "percentage",
  time.var,
  event.var,
  cutoff.type = "maxstat",
  variables
)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT:fcs.SCE]{FlowCT::fcs.SCE()}}.}

\item{cell.clusters}{Name of column containing clusters identified through \code{\link[FlowCT:clustering.flow]{FlowCT::clustering.flow()}}.}

\item{value}{String specifying if final resuls should be proportions ("percentage", default) or raw counts ("counts").}

\item{time.var}{Survival time variable.}

\item{event.var}{Variable with event censoring.}

\item{cutoff.type}{Method for calculating survival cutoffs. Available methods are "maxstat" (default, based on \href{https://cran.r-project.org/web/packages/maxstat/index.html}{\code{maxstat}}), "quantiles" (i.e., terciles) and "median".}

\item{variables}{Vector with variables for calculating the cutoff. If nothing is detailed (\code{NULL}, default), all immune populations from \code{cell.clusters} will be considered.}
}
\description{
Once immune cell populations are identified, it calculates a cutoff from the percentage (or raw counts) according survival time.
}
\examples{
\dontrun{
ct <- pop.cutoff(fcs.SCE = fcs, cell.clusters = "SOM_named", time.var = "PFS",
    event.var = "PFS_c", cutoff.type = "quantiles")
}
}
\keyword{cutoffs}
\keyword{survival}
