% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{qc.and.removeDoublets}
\alias{qc.and.removeDoublets}
\title{qc.and.removeDoublets}
\usage{
qc.and.removeDoublets(fcs.SE = NULL, filelist = NULL,
  directory = getwd(), pattern = "fcs", physical.markers,
  output.folder = NULL, output.suffix = "qc", return.idx = F)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}. If \code{NULL} (default), this function will search for FCS files within current working directory.}

\item{filelist}{A vector with full path of FCS files to be read, commonly generated through \code{\link[base:list.files]{base::list.files()}}. If \code{NULL}, this file list will be generated as indicated below.}

\item{directory}{If \code{filelist = NULL}, those files stored in this location will be read. Default = \code{getwd()} (current directory).}

\item{pattern}{Pattern for reading files within \code{directory}. Default = \code{"fcs"}.}

\item{physical.markers}{Vector with physical markers (i.e., FCS-A, SSC-H, etc.). Mandatory for perfoming quality control.}

\item{output.folder}{Folder name for storing final FCS files. Default = \code{NULL} (current directory).}

\item{output.suffix}{Suffix added to new generated FCS files (in case not being working with a \code{FCS.SE} object). Default = \code{qc}.}

\item{return.idx}{Logical indicating whether output is a new \code{FCS.SE} object without low-quality events or an index with this positions (without removing them from original \code{FCS.SE}). This option is only available if the input is a \code{FCS.SE} object. Default = \code{F} (\code{FCS.SE} output).}
}
\value{
Final output is a new \code{FCS.SE} object (if FCS.SE input) without these events that do not pass the quality control or new FCS files with the suffix \code{qc} with these low quality events removed. If user specifies \code{return.idx = T}, output would be a vector with all low-quality events positions.

A table with percentaje of removed events for each FCS will be also shown in the terminal (those files with more than 30% of events removed will have a '!' signal).
}
\description{
qc.and.removeDoublets
}
\examples{
\dontrun{
# option 1: output a new FCS.SE object with low-quality events removed.
fcs_se_qc <- qc.and.removeDoublets(fcs.SE = fcs_se, 
    physical.markers = c("FSC_A", "FSC_H", "SSC_A", "SSC_H"))

# option 2: output an index with low quality events
idx_qc <- qc.and.removeDoublets(fcs.SE = fcs_se, 
    physical.markers = c("FSC_A", "FSC_H", "SSC_A", "SSC_H"), return.idx = T)

# option 3: working with all FCS files within a folder
idx_qc <- qc.and.removeDoublets(directory = "../data/", 
    physical.markers = c("FSC_A", "FSC_H", "SSC_A", "SSC_H"), output.folder = "HQ_files")
}
}
\keyword{control}
\keyword{doublets}
\keyword{events}
\keyword{low}
\keyword{quality}
\keyword{removal}
\keyword{remove}
