% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.and.removeDoublets.R
\name{qc.and.removeDoublets}
\alias{qc.and.removeDoublets}
\title{qc.and.removeDoublets}
\usage{
qc.and.removeDoublets(fcs.SCE = NULL, filelist = NULL,
  directory = getwd(), pattern = "fcs", physical.markers,
  output.folder = directory, output.suffix = "qc",
  return.idx = FALSE, return.fcs = FALSE)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT]{fcs.SCE()}}. If \code{NULL} (default), this function will search for FCS files within current working directory.}

\item{filelist}{A vector with full path of FCS files to be read, commonly generated through \code{\link[base]{list.files()}}. If \code{NULL}, this file list will be generated as indicated below.}

\item{directory}{If \code{filelist = NULL}, those files stored in this location will be read. Default = \code{getwd()} (current directory).}

\item{pattern}{Pattern for reading files within \code{directory}. Default = \code{"fcs"}.}

\item{physical.markers}{Vector with physical markers (i.e., FCS-A, SSC-H, etc.). Mandatory for perfoming quality control.}

\item{output.folder}{Folder name for storing final FCS files. Default, the same of \code{directory}.}

\item{output.suffix}{Suffix added to new generated FCS files (in case not being working with a \code{fcs.SCE} object). Default = \code{qc}.}

\item{return.idx}{Logical indicating whether output is a new \code{fcs.SCE} object without low-quality events or an index with this positions (without removing them from original \code{fcs.SCE}). This option is only available if the input is a \code{fcs.SCE} object. Default = \code{FALSE} (\code{fcs.SCE} output).}

\item{return.fcs}{Logical indicating whether new FCS file should be generated without low quality events. They will stored in specified \code{output.folder} with the corresponding \code{output.suffix}.}
}
\value{
Final output is a new \code{fcs.SCE} object (if \code{fcs.SCE} input) without these events that do not pass the quality control or new FCS files with the suffix \code{"qc"} with these low quality events removed. If user specifies \code{return.idx = TRUE}, output would be a vector with all low-quality events positions.

A table with percentaje of removed events for each FCS will be also shown in the terminal (those files with more than 30% of events removed will have a '!' signal).
}
\description{
qc.and.removeDoublets
}
\examples{
\dontrun{
# option 1: output a new fcs.SCE object with low-quality events removed.
fcs.SCE_qc <- qc.and.removeDoublets(fcs.SCE = fcs, 
     physical.markers = c("FSC_A", "FSC_H", "SSC_A", "SSC_H"))

# option 2: output an index with low quality events
idx_qc <- qc.and.removeDoublets(fcs.SCE = fcs, 
     physical.markers = c("FSC_A", "FSC_H", "SSC_A", "SSC_H"), return.idx = T)

# option 3: working with all FCS files within a folder
idx_qc <- qc.and.removeDoublets(directory = "../data/", output.folder = "HQ_files",
     physical.markers = c("FSC_A", "FSC_H", "SSC_A", "SSC_H"))
}
}
\keyword{control}
\keyword{doublets}
\keyword{events}
\keyword{low}
\keyword{quality}
\keyword{removal}
\keyword{remove}
