% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{qc.and.removeDoublets}
\alias{qc.and.removeDoublets}
\title{qc.and.removeDoublets}
\usage{
qc.and.removeDoublets(directory = ".", reduction_computed = TRUE,
  reduction_suffix = "_red", output_suffix = "_preprocessed",
  output_folder = "results_HQsinglets")
}
\arguments{
\item{directory}{Directory with FCS files to process. Default = "."}

\item{reduction_computed}{Logical whether reduction previously computed with \code{\link[FlowCT:reduce.FCS]{FlowCT::reduce.FCS()}}. Default = \code{TRUE}}

\item{reduction_suffix}{Suffix to consider in reduced files (normaly it would be "_red" if it is coming from \code{\link[FlowCT:reduce.FCS]{FlowCT::reduce.FCS()}}). Default = "_red"}

\item{output_suffix}{Suffix to be added to final FCS files. Default = "_preprocessed"}

\item{output_folder}{Folder name for storing final FCS files. Default = "results_HQsinglets"}
}
\description{
This function performs a double-step procedure in all FCS files cointained within a folder:\enumerate{
   \item It executes a quality control based in \href{http://bioconductor.org/packages/release/bioc/vignettes/flowAI/inst/doc/flowAI.html}{\code{\link[flowAI:flow_auto_qc]{flowAI::flow_auto_qc()}}},
   \item Remove doublets according criteria described \href{https://github.com/LennonLab/flow-cytometry/blob/fcf09fc7b4943a386864de9b8ee43ce4e1c8d34d/bin/support_functions.R}{here}}.
}
\examples{
\dontrun{qc.and.removeDoublets()}
}
\keyword{control}
\keyword{doublets}
\keyword{quality}
\keyword{remove}
