% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.pop.R
\name{remove.pop}
\alias{remove.pop}
\title{It removes one or multiple cell populations from a \code{fcs.SCE} object.}
\usage{
remove.pop(fcs.SCE, population, clusters.named)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT]{fcs.SCE()}}.}

\item{population}{Name(s) of cell population(s) to be removed.}

\item{clusters.named}{Column name from the \code{colData(fcs.SCE)} object which contains renamed clusters (through \code{\link[FlowCT]{clusters.rename()}}).}
}
\description{
It removes one or multiple cell populations from a \code{fcs.SCE} object.
}
\examples{
\dontrun{
fcs_rm <- remove.pop(fcs, clusters.named = "SOM_named", 
    population = c("debris", "unclassified"))
}
}
\keyword{cell}
\keyword{debris}
\keyword{population}
\keyword{remove}
