% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sub.samples}
\alias{sub.samples}
\title{sub.samples}
\usage{
sub.samples(fcs.SE, subsampling = 1000, index = F)
}
\arguments{
\item{fcs.SE}{A FCS.SE object generated through \code{\link[FlowCT:fcs.SE]{FlowCT::fcs.SE()}}.}

\item{subsampling}{Number of events to keep. Default = 1000.}

\item{index}{Logical indicating if returning an FCS.SE object or a index vector. Default = \code{FALSE}.}
}
\description{
It generates reduce randomly the number of events of a FCS.SE object (it computes this reduction for each FCS file separatelly inside this object). It can generate a new reduced FCS.SE object or a simple index position for removing events.
}
\examples{
\dontrun{
# option 1
fcs_se_red <- sub.samples(data = fcs_se, subsampling = 1000)

# option 2
keep_red <- sub.samples(data = fcs_se, subsampling = 1000, index = T)
fcs_se_red <- fcs_se[,keep_red]
}
}
\keyword{subsampling}
