% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.samplesR.R
\name{sub.samples}
\alias{sub.samples}
\title{sub.samples}
\usage{
sub.samples(fcs.SCE, subsampling = 1000, index = F)
}
\arguments{
\item{fcs.SCE}{A \code{fcs.SCE} object generated through \code{\link[FlowCT:fcs.SCE]{fcs.SCE()}}.}

\item{subsampling}{Number of events to keep. Default = 1000.}

\item{index}{Logical indicating if returning an fcs.SCE object or a index vector. Default = \code{FALSE}.}
}
\description{
It generates reduce randomly the number of events of a fcs.SCE object (it computes this reduction for each FCS file separatelly inside this object). It can generate a new reduced \code{fcs.SCE} object or a simple index position for removing events.
}
\examples{
\dontrun{
# option 1
fcs_red <- sub.samples(data = fcs, subsampling = 1000)

# option 2
keep_red <- sub.samples(data = fcs, subsampling = 1000, index = T)
fcs_red <- fcs[,keep_red]
}
}
\keyword{subsampling}
